@@ -12,18 +12,10 @@ namespace Server.Controllers.Case.Personal;
 [ApiController]
 [Route("/case/{caseId}")]
 [AuditApi(EventTypeName = "HTTP")]
-public class ContemporaneousNotesController : ControllerBase
+public class ContemporaneousNotesController(DatabaseContext dbContext, SqidsEncoder<long> sqids, IConfiguration configuration)
+    : ControllerBase
 {
-    private readonly AuditScopeFactory _auditContext;
-    private readonly DatabaseContext _dbContext;
-    private readonly SqidsEncoder<long> _sqids;
-
-    public ContemporaneousNotesController(DatabaseContext dbContext, SqidsEncoder<long> sqids)
-    {
-        _dbContext = dbContext;
-        _auditContext = new AuditScopeFactory();
-        _sqids = sqids;
-    }
+    private readonly AuditScopeFactory _auditContext = new();
 
     // GET:  /case/?/contemporaneous-notes
     [HttpGet("contemporaneous-notes")]
@@ -32,39 +24,37 @@ public class ContemporaneousNotesController : ControllerBase
     [ProducesResponseType(typeof(string), StatusCodes.Status401Unauthorized)]
     [ProducesResponseType(typeof(string), StatusCodes.Status404NotFound)]
     [ProducesResponseType(typeof(string), StatusCodes.Status500InternalServerError)]
-    [Authorize(Roles = "user")]
+    [Authorize]
     public async Task<ActionResult<List<API.ContemporaneousNotes>>> GetContemporaneousNotes(string caseId)
     {
         // Preflight checks
         PreflightResponse preflightResponse = await PreflightChecks();
 
-        // If preflight checks returned a HTTP error raise it here
+        // If preflight checks returned an HTTP error raise it here
         if (preflightResponse.Error != null) return preflightResponse.Error;
 
-        // If preflight checks details are null return a HTTP 500 error
+        // If preflight checks details are null return an HTTP 500 error
         if (preflightResponse.Details == null)
             return Problem(
                 "Preflight checks failed with an unknown error!"
             );
 
         // Set variables from preflight response
-        string organizationId = preflightResponse.Details.OrganizationId;
-        long userId = preflightResponse.Details.UserId;
+        string emailAddress = preflightResponse.Details.EmailAddress;
         Database.UserSettings userSettings = preflightResponse.Details.UserSettings;
-        long rawCaseId = _sqids.Decode(caseId)[0];
+        long rawCaseId = sqids.Decode(caseId)[0];
 
-        // Log the user's organization ID and the user's ID
+        // Log the user's email address
         IAuditScope auditScope = this.GetCurrentAuditScope();
-        auditScope.SetCustomField("OrganizationID", organizationId);
-        auditScope.SetCustomField("UserID", userId);
+        auditScope.SetCustomField("emailAddress", emailAddress);
 
-        // Get case user from the database including the required entities 
-        Database.CaseUser? caseUser = await _dbContext.CaseUser
-            .Where(cu => cu.Case.Id == rawCaseId && cu.User.Id == userId)
+        // Get case user from the database including the required entities
+        Database.CaseUser? caseUser = await dbContext.CaseUser
+            .Where(cu => cu.Case.Id == rawCaseId && cu.User.EmailAddress == emailAddress)
             .Include(cu => cu.ContemporaneousNotes)
             .SingleOrDefaultAsync();
 
-        // If case user does not exist then return a HTTP 404 error 
+        // If case user does not exist then return an HTTP 404 error
         if (caseUser == null) return NotFound($"The case `{caseId}` does not exist!");
         // The case might not exist or the user does not have access to the case
 
@@ -74,84 +64,81 @@ public class ContemporaneousNotesController : ControllerBase
         // Return a list of the user's contemporaneous notes
         return caseUser.ContemporaneousNotes.Select(cn => new API.ContemporaneousNotes
         {
-            Id = _sqids.Encode(cn.Id),
-            Created = TimeZoneInfo.ConvertTimeFromUtc(cn.Created, timeZone)
+            Id = sqids.Encode(cn.Id), Created = TimeZoneInfo.ConvertTimeFromUtc(cn.Created, timeZone)
         }).ToList();
     }
 
     // GET: /case/?/contemporaneous-note/?
-    [HttpGet("contemporaneous-note/{noteId}")]
+    [HttpGet("contemporaneous-note/{noteEmailAddress}")]
     [ProducesResponseType(StatusCodes.Status200OK)]
     [ProducesResponseType(typeof(string), StatusCodes.Status400BadRequest)]
     [ProducesResponseType(typeof(string), StatusCodes.Status401Unauthorized)]
     [ProducesResponseType(typeof(string), StatusCodes.Status404NotFound)]
     [ProducesResponseType(typeof(string), StatusCodes.Status500InternalServerError)]
-    [Authorize(Roles = "user")]
-    public async Task<ActionResult> GetContemporaneousNote(string caseId, string noteId)
+    [Authorize]
+    public async Task<ActionResult> GetContemporaneousNote(string caseId, string noteEmailAddress)
     {
         // Preflight checks
         PreflightResponse preflightResponse = await PreflightChecks();
 
-        // If preflight checks returned a HTTP error raise it here
+        // If preflight checks returned an HTTP error raise it here
         if (preflightResponse.Error != null) return preflightResponse.Error;
 
-        // If preflight checks details are null return a HTTP 500 error
+        // If preflight checks details are null return an HTTP 500 error
         if (preflightResponse.Details == null)
             return Problem(
                 "Preflight checks failed with an unknown error!"
             );
 
         // Set variables from preflight response
-        string organizationId = preflightResponse.Details.OrganizationId;
-        Database.OrganizationSettings organizationSettings = preflightResponse.Details.OrganizationSettings;
-        long userId = preflightResponse.Details.UserId;
-        long rawCaseId = _sqids.Decode(caseId)[0];
+        string emailAddress = preflightResponse.Details.EmailAddress;
+        long rawCaseId = sqids.Decode(caseId)[0];
 
-        // Log the user's organization ID and the user's ID
+        // Log the user's ID
         IAuditScope auditScope = this.GetCurrentAuditScope();
-        auditScope.SetCustomField("OrganizationID", organizationId);
-        auditScope.SetCustomField("UserID", userId);
 
-        // Get case user from the database including the required entities 
-        Database.CaseUser? caseUser = await _dbContext.CaseUser
-            .Where(cu => cu.Case.Id == rawCaseId && cu.User.Id == userId)
+        auditScope.SetCustomField("emailAddress", emailAddress);
+
+        // Get case user from the database including the required entities
+        Database.CaseUser? caseUser = await dbContext.CaseUser
+            .Where(cu => cu.Case.Id == rawCaseId && cu.User.EmailAddress == emailAddress)
             .Include(cu => cu.ContemporaneousNotes)
             .Include(cu => cu.Hashes)
             .SingleOrDefaultAsync();
 
-        // If case user does not exist then return a HTTP 404 error 
+        // If case user does not exist then return an HTTP 404 error
         if (caseUser == null) return NotFound($"The case `{caseId}` does not exist!");
         // The case might not exist or the user does not have access to the case
 
-        // Convert note ID squid to ID 
-        long rawNoteId = _sqids.Decode(noteId)[0];
+        // Convert note ID squid to ID
+        long rawNoteId = sqids.Decode(noteEmailAddress)[0];
 
         // Fetch the contemporaneous note from the database
         Database.ContemporaneousNote? contemporaneousNote =
             caseUser.ContemporaneousNotes.SingleOrDefault(cn => cn.Id == rawNoteId);
 
-        // If contemporaneous note is null then return a HTTP 404 error as it does not exist
-        if (contemporaneousNote == null) return NotFound($"Can not find the note with the ID `{noteId}`");
+        // If contemporaneous note is null then return an HTTP 404 error as it does not exist
+        if (contemporaneousNote == null) return NotFound($"Can not find the note with the ID `{noteEmailAddress}`");
 
         // Create minio client
         IMinioClient minio = new MinioClient()
-            .WithEndpoint(organizationSettings.S3Endpoint)
-            .WithCredentials(organizationSettings.S3AccessKey, organizationSettings.S3SecretKey)
-            .WithSSL(organizationSettings.S3NetworkEncryption)
+            .WithEndpoint(configuration.GetValue<string>("Minio:Endpoint"))
+            .WithCredentials(configuration.GetValue<string>("Minio:AccessKey"), configuration.GetValue<string>("Minio:SecretKey"))
+            .WithSSL(configuration.GetValue<bool>("Minio:NetworkEncryption"))
             .Build();
 
-        // Create a variable for object path with auth0| removed 
+        // Create a variable for object path
         string objectPath =
-            $"cases/{caseId}/{_sqids.Encode(userId)}/contemporaneous-notes/{noteId}/note.txt";
+            $"cases/{caseId}/{emailAddress}/contemporaneous-notes/{noteEmailAddress}/note.txt";
 
         // Check if bucket exists
         bool bucketExists = await minio.BucketExistsAsync(new BucketExistsArgs()
-            .WithBucket(organizationSettings.S3BucketName)
+            .WithBucket(configuration.GetValue<string>("Minio:BucketName"))
         ).ConfigureAwait(false);
 
         // If bucket does not exist return HTTP a 500 error
         if (!bucketExists)
-            return Problem($"An S3 Bucket with the name `{organizationSettings.S3BucketName}` does not exist!");
+            return Problem($"An S3 Bucket with the name `{configuration.GetValue<string>("Minio:BucketName")}` does not exist!");
 
         // Create variable to store object metadata
         ObjectStat objectMetadata;
@@ -160,7 +147,7 @@ public class ContemporaneousNotesController : ControllerBase
         try
         {
             objectMetadata = await minio.StatObjectAsync(new StatObjectArgs()
-                .WithBucket(organizationSettings.S3BucketName)
+                .WithBucket(configuration.GetValue<string>("Minio:BucketName"))
                 .WithObject(objectPath)
             );
         }
@@ -168,17 +155,17 @@ public class ContemporaneousNotesController : ControllerBase
         catch (ObjectNotFoundException)
         {
             return Problem(
-                $"Can not find the S3 object for contemporaneous note: `{noteId}` at the following path: `{objectPath}`.");
+                $"Can not find the S3 object for contemporaneous note: `{noteEmailAddress}` at the following path: `{objectPath}`.");
         }
 
         // Fetch object hash from database
         Database.Hash? objectHashes = caseUser.Hashes.SingleOrDefault(h =>
             h.ObjectName == objectMetadata.ObjectName && h.VersionId == objectMetadata.VersionId);
 
-        // If object hash is null then a hash does not exist so return a HTTP 500 error
+        // If object hash is null then a hash does not exist so return an HTTP 500 error
         if (objectHashes == null)
             return Problem(
-                $"Unable to find hash value for contemporaneous note with the ID `{_sqids.Encode(contemporaneousNote.Id)}` at the path `{objectPath}`!",
+                $"Unable to find hash value for contemporaneous note with the ID `{sqids.Encode(contemporaneousNote.Id)}` at the path `{objectPath}`!",
                 title: "Could not find hash value for contemporaneous note!");
 
         // Create memory stream to store file contents
@@ -186,7 +173,7 @@ public class ContemporaneousNotesController : ControllerBase
 
         // Get object and copy file contents to stream
         await minio.GetObjectAsync(new GetObjectArgs()
-            .WithBucket(organizationSettings.S3BucketName)
+            .WithBucket(configuration.GetValue<string>("Minio:BucketName"))
             .WithObject(objectPath)
             .WithCallbackStream(stream => { stream.CopyTo(memoryStream); })
         );
@@ -205,13 +192,13 @@ public class ContemporaneousNotesController : ControllerBase
         // Check generated MD5 hash matches the hash in the database
         if (BitConverter.ToString(md5Hash).Replace("-", "").ToLowerInvariant() != objectHashes.Md5Hash)
             return Problem(
-                $"MD5 hash verification failed for contemporaneous note with the ID `{_sqids.Encode(contemporaneousNote.Id)}` at the path `{objectPath}`!",
+                $"MD5 hash verification failed for contemporaneous note with the ID `{sqids.Encode(contemporaneousNote.Id)}` at the path `{objectPath}`!",
                 title: "MD5 hash verification failed!");
 
         // Check generated SHA256 hash matches the hash in the database
         if (BitConverter.ToString(sha256Hash).Replace("-", "").ToLowerInvariant() != objectHashes.ShaHash)
             return Problem(
-                $"SHA256 hash verification failed for contemporaneous note with the ID `{_sqids.Encode(contemporaneousNote.Id)}` at the path `{objectPath}`!",
+                $"SHA256 hash verification failed for contemporaneous note with the ID `{sqids.Encode(contemporaneousNote.Id)}` at the path `{objectPath}`!",
                 title: "SHA256 hash verification failed!");
 
         // Return file
@@ -225,73 +212,70 @@ public class ContemporaneousNotesController : ControllerBase
     [ProducesResponseType(typeof(string), StatusCodes.Status401Unauthorized)]
     [ProducesResponseType(typeof(string), StatusCodes.Status404NotFound)]
     [ProducesResponseType(typeof(string), StatusCodes.Status500InternalServerError)]
-    [Authorize(Roles = "user")]
-    public async Task<ActionResult> PostContemporaneousNote(string caseId, [FromForm] IFormFile file)
+    [Authorize]
+    public async Task<ActionResult> PostContemporaneousNote(string caseId, IFormFile file)
     {
         // Preflight checks
         PreflightResponse preflightResponse = await PreflightChecks();
 
-        // If preflight checks returned a HTTP error raise it here
+        // If preflight checks returned an HTTP error raise it here
         if (preflightResponse.Error != null) return preflightResponse.Error;
 
-        // If preflight checks details are null return a HTTP 500 error
+        // If preflight checks details are null return an HTTP 500 error
         if (preflightResponse.Details == null)
             return Problem(
                 "Preflight checks failed with an unknown error!"
             );
 
         // Set variables from preflight response
-        string organizationId = preflightResponse.Details.OrganizationId;
-        Database.OrganizationSettings organizationSettings = preflightResponse.Details.OrganizationSettings;
-        long userId = preflightResponse.Details.UserId;
+        string emailAddress = preflightResponse.Details.EmailAddress;
         string userNameJob = preflightResponse.Details.UserNameJob;
-        long rawCaseId = _sqids.Decode(caseId)[0];
+        long rawCaseId = sqids.Decode(caseId)[0];
 
-        // Log the user's organization ID and the user's ID
+        // Log the user's email address
         IAuditScope auditScope = this.GetCurrentAuditScope();
-        auditScope.SetCustomField("OrganizationID", organizationId);
-        auditScope.SetCustomField("UserID", userId);
+        auditScope.SetCustomField("emailAddress", emailAddress);
 
-        // Get case user from the database including the required entities 
-        Database.CaseUser? caseUser = await _dbContext.CaseUser
-            .Where(cu => cu.Case.Id == rawCaseId && cu.User.Id == userId)
+        // Get case user from the database including the required entities
+        Database.CaseUser? caseUser = await dbContext.CaseUser
+            .Where(cu => cu.Case.Id == rawCaseId && cu.User.EmailAddress == emailAddress)
             .Include(cu => cu.ContemporaneousNotes)
             .Include(cu => cu.Hashes)
             .Include(cu => cu.Case)
             .SingleOrDefaultAsync();
 
-        // If case user does not exist then return a HTTP 404 error 
+        // If case user does not exist then return an HTTP 404 error
         if (caseUser == null) return NotFound($"The case `{caseId}` does not exist!");
         // The case might not exist or the user does not have access to the case
 
         // Create minio client
         IMinioClient minio = new MinioClient()
-            .WithEndpoint(organizationSettings.S3Endpoint)
-            .WithCredentials(organizationSettings.S3AccessKey, organizationSettings.S3SecretKey)
-            .WithSSL(organizationSettings.S3NetworkEncryption)
+            .WithEndpoint(configuration.GetValue<string>("Minio:Endpoint"))
+            .WithCredentials(configuration.GetValue<string>("Minio:AccessKey"), configuration.GetValue<string>("Minio:SecretKey"))
+            .WithSSL(configuration.GetValue<bool>("Minio:NetworkEncryption"))
             .Build();
 
         // Create contemporaneous note record in the database
         Database.ContemporaneousNote contemporaneousNote = new();
 
-        // Add the the note to the collection
+        // Add the note to the collection
         caseUser.ContemporaneousNotes.Add(contemporaneousNote);
 
         // Save changes to the database so a GUID is generated
-        await _dbContext.SaveChangesAsync();
+        await dbContext.SaveChangesAsync();
 
-        // Create a variable for object path with auth0| removed 
+        // Create a variable for object path
         string objectPath =
-            $"cases/{caseId}/{_sqids.Encode(userId)}/contemporaneous-notes/{_sqids.Encode(contemporaneousNote.Id)}/note.txt";
+            $"cases/{caseId}/{emailAddress}/contemporaneous-notes/{sqids.Encode(contemporaneousNote.Id)}/note.txt";
 
         // Check if bucket exists
         bool bucketExists = await minio.BucketExistsAsync(new BucketExistsArgs()
-            .WithBucket(organizationSettings.S3BucketName)
+            .WithBucket(configuration.GetValue<string>("Minio:BucketName"))
         ).ConfigureAwait(false);
 
-        // If bucket does not exist return a HTTP 500 error
+        // If bucket does not exist return an HTTP 500 error
         if (!bucketExists)
-            return Problem($"An S3 Bucket with the name `{organizationSettings.S3BucketName}` does not exist!");
+            return Problem($"An S3 Bucket with the name `{configuration.GetValue<string>("Minio:BucketName")}` does not exist!");
 
         try
         {
@@ -305,7 +289,7 @@ public class ContemporaneousNotesController : ControllerBase
 
             // Save file to s3 bucket
             await minio.PutObjectAsync(new PutObjectArgs()
-                .WithBucket(organizationSettings.S3BucketName)
+                .WithBucket(configuration.GetValue<string>("Minio:BucketName"))
                 .WithObject(objectPath)
                 .WithStreamData(memoryStream)
                 .WithObjectSize(memoryStream.Length)
@@ -329,7 +313,7 @@ public class ContemporaneousNotesController : ControllerBase
 
             // Create Meilisearch Client
             MeilisearchClient meiliClient =
-                new(organizationSettings.MeilisearchUrl, organizationSettings.MeilisearchApiKey);
+                new(configuration.GetValue<string>("Meilisearch:Url"), configuration.GetValue<string>("Meilisearch:Key"));
 
             // Try getting the contemporaneous-notes index
             try
@@ -344,30 +328,26 @@ public class ContemporaneousNotesController : ControllerBase
                 {
                     await meiliClient.CreateIndexAsync("contemporaneous-notes", "id");
                     await meiliClient.Index("contemporaneous-notes")
-                        .UpdateFilterableAttributesAsync(new[] { "caseId", "userId" });
+                        .UpdateFilterableAttributesAsync(["caseId", "emailAddress"]);
                 }
             }
 
             // Get the contemporaneous-notes index
             Index index = meiliClient.Index("contemporaneous-notes");
 
-            await index.AddDocumentsAsync(new[]
-            {
+            await index.AddDocumentsAsync([
                 new Search.ContemporaneousNote
                 {
-                    Id = contemporaneousNote.Id,
-                    UserId = userId,
-                    CaseId = _sqids.Decode(caseId)[0],
-                    Content = content
+                    Id = contemporaneousNote.Id, EmailAddress = emailAddress, CaseId = sqids.Decode(caseId)[0], Content = content
                 }
-            });
+            ]);
 
             // Set memory stream position to 0 as per github.com/minio/minio/issues/6274
             memoryStream.Position = 0;
 
             // Fetch object metadata
             ObjectStat objectMetadata = await minio.StatObjectAsync(new StatObjectArgs()
-                .WithBucket(organizationSettings.S3BucketName)
+                .WithBucket(configuration.GetValue<string>("Minio:BucketName"))
                 .WithObject(objectPath)
             );
 
@@ -389,15 +369,15 @@ public class ContemporaneousNotesController : ControllerBase
             });
 
             // Save changes to the database
-            await _dbContext.SaveChangesAsync();
+            await dbContext.SaveChangesAsync();
 
             // Log the creation
             await _auditContext.LogAsync("Lighthouse Notes",
                 new
                 {
                     Action =
-                        $"`{userNameJob}` created contemporaneous note for `{caseUser.Case.DisplayName}` with the ID `{_sqids.Encode(contemporaneousNote.Id)}`.",
-                    UserID = userId, OrganizationID = organizationId
+                        $"`{userNameJob}` created contemporaneous note for `{caseUser.Case.DisplayName}` with the ID `{sqids.Encode(contemporaneousNote.Id)}`.",
+                    EmailAddress = emailAddress
                 });
 
             // Return Ok
@@ -417,47 +397,44 @@ public class ContemporaneousNotesController : ControllerBase
     [ProducesResponseType(typeof(string), StatusCodes.Status401Unauthorized)]
     [ProducesResponseType(typeof(string), StatusCodes.Status404NotFound)]
     [ProducesResponseType(typeof(string), StatusCodes.Status500InternalServerError)]
-    [Authorize(Roles = "user")]
+    [Authorize]
     public async Task<ActionResult<List<API.ContemporaneousNotes>>> GetSearchContemporaneousNotes(string caseId,
         API.Search search)
     {
         // Preflight checks
         PreflightResponse preflightResponse = await PreflightChecks();
 
-        // If preflight checks returned a HTTP error raise it here
+        // If preflight checks returned an HTTP error raise it here
         if (preflightResponse.Error != null) return preflightResponse.Error;
 
-        // If preflight checks details are null return a HTTP 500 error
+        // If preflight checks details are null return an HTTP 500 error
         if (preflightResponse.Details == null)
             return Problem(
                 "Preflight checks failed with an unknown error!"
             );
 
         // Set variables from preflight response
-        string organizationId = preflightResponse.Details.OrganizationId;
-        Database.OrganizationSettings organizationSettings = preflightResponse.Details.OrganizationSettings;
-        long userId = preflightResponse.Details.UserId;
-        long rawCaseId = _sqids.Decode(caseId)[0];
+        string emailAddress = preflightResponse.Details.EmailAddress;
+        long rawCaseId = sqids.Decode(caseId)[0];
         Database.UserSettings userSettings = preflightResponse.Details.UserSettings;
 
-        // Log the user's organization ID and the user's ID
+        // Log the user's email address
         IAuditScope auditScope = this.GetCurrentAuditScope();
-        auditScope.SetCustomField("OrganizationID", organizationId);
-        auditScope.SetCustomField("UserID", userId);
+        auditScope.SetCustomField("emailAddress", emailAddress);
 
-        // Get case user from the database including the required entities 
-        Database.CaseUser? caseUser = await _dbContext.CaseUser
-            .Where(cu => cu.Case.Id == rawCaseId && cu.User.Id == userId)
+        // Get case user from the database including the required entities
+        Database.CaseUser? caseUser = await dbContext.CaseUser
+            .Where(cu => cu.Case.Id == rawCaseId && cu.User.EmailAddress == emailAddress)
             .Include(cu => cu.ContemporaneousNotes)
             .SingleOrDefaultAsync();
 
-        // If case user does not exist then return a HTTP 404 error 
+        // If case user does not exist then return an HTTP 404 error
         if (caseUser == null) return NotFound($"The case `{caseId}` does not exist!");
         // The case might not exist or the user does not have access to the case
 
         // Create Meilisearch Client
         MeilisearchClient meiliClient =
-            new(organizationSettings.MeilisearchUrl, organizationSettings.MeilisearchApiKey);
+            new(configuration.GetValue<string>("Meilisearch:Url"), configuration.GetValue<string>("Meilisearch:Key"));
 
         // Get the contemporaneous-notes index
         Index index = meiliClient.Index("contemporaneous-notes");
@@ -467,13 +444,13 @@ public class ContemporaneousNotesController : ControllerBase
             search.Query,
             new SearchQuery
             {
-                AttributesToSearchOn = new[] { "content" },
-                AttributesToRetrieve = new[] { "id" },
-                Filter = $"caseId = {rawCaseId} AND userId = {userId}"
+                AttributesToSearchOn = ["content"],
+                AttributesToRetrieve = ["id"],
+                Filter = $"caseId = {rawCaseId} AND emailAddress = {emailAddress}"
             }
         );
 
-        // Create a list of contemporaneous notes Ids that contain a match 
+        // Create a list of contemporaneous notes ids that contain a match
         List<long> contemporaneousNotesIds = searchResult.Hits.Select(cn => cn.Id).ToList();
 
         // Get the user's time zone
@@ -483,53 +460,36 @@ public class ContemporaneousNotesController : ControllerBase
         return caseUser.ContemporaneousNotes.Where(cn => contemporaneousNotesIds.Contains(cn.Id))
             .Select(cn => new API.ContemporaneousNotes
             {
-                Id = _sqids.Encode(cn.Id),
-                Created = TimeZoneInfo.ConvertTimeFromUtc(cn.Created, timeZone)
+                Id = sqids.Encode(cn.Id), Created = TimeZoneInfo.ConvertTimeFromUtc(cn.Created, timeZone)
             }).ToList();
     }
 
     private async Task<PreflightResponse> PreflightChecks()
     {
-        // Get user ID from claim
-        string? auth0UserId = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value;
+        // Get user email from JWT claim
+        string? emailAddress = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Email)?.Value;
 
-        // If user ID is null then it does not exist in JWT so return a HTTP 400 error
-        if (auth0UserId == null)
-            return new PreflightResponse
-                { Error = BadRequest("User ID can not be found in the JSON Web Token (JWT)!") };
+        // If user email is null then it does not exist in JWT so return an HTTP 400 error
+        if (string.IsNullOrEmpty(emailAddress))
+            return new PreflightResponse { Error = BadRequest("Email Address cannot be found in the JSON Web Token (JWT)!") };
 
-        // Get organization ID from claim
-        string? organizationId = User.Claims.FirstOrDefault(c => c.Type == "org_id")?.Value;
-
-        // If organization ID  is null then it does not exist in JWT so return a HTTP 400 error
-        if (organizationId == null)
-            return new PreflightResponse
-                { Error = BadRequest("Organization ID can not be found in the JSON Web Token (JWT)!") };
-
-        // Select organization ID, organization settings, user ID and user name and job from the user table
-        PreflightResponseDetails? userQueryResult = await _dbContext.User
-            .Where(u => u.Auth0Id == auth0UserId && u.Organization.Id == organizationId)
+        // Query user details including roles and application settings
+        PreflightResponseDetails? preflightData = await dbContext.User
+            .Where(u => u.EmailAddress == emailAddress)
+            .Include(u => u.Settings)
             .Select(u => new PreflightResponseDetails
             {
-                OrganizationId = u.Organization.Id,
-                OrganizationSettings = u.Organization.Settings,
-                UserId = u.Id,
-                UserNameJob = $"{u.DisplayName} ({u.JobTitle})",
-                UserSettings = u.Settings
-            }).SingleOrDefaultAsync();
+                EmailAddress = u.EmailAddress, UserNameJob = $"{u.DisplayName} ({u.JobTitle})", UserSettings = u.Settings
+            })
+            .SingleOrDefaultAsync();
 
-        // If query result is null then the user does not exit in the organization so return a HTTP 404 error
-        if (userQueryResult == null)
-            return new PreflightResponse
-            {
-                Error = NotFound(
-                    $"A user with the Auth0 user ID `{auth0UserId}` was not found in the organization with the Auth0 organization ID `{organizationId}`!")
-            };
+        // If query result is null then the user does not exist
+        if (preflightData == null)
+            return new PreflightResponse { Error = NotFound($"A user with the user email: `{emailAddress}` was not found!") }
+                ;
 
-        return new PreflightResponse
-        {
-            Details = userQueryResult
-        };
+        // Return preflight response
+        return new PreflightResponse { Details = preflightData };
     }
 
     private class PreflightResponse
@@ -540,9 +500,7 @@ public class ContemporaneousNotesController : ControllerBase
 
     private class PreflightResponseDetails
     {
-        public required string OrganizationId { get; init; }
-        public required Database.OrganizationSettings OrganizationSettings { get; init; }
-        public long UserId { get; init; }
+        public required string EmailAddress { get; init; }
         public required string UserNameJob { get; init; }
         public required Database.UserSettings UserSettings { get; init; }
     }
