@@ -9,18 +9,9 @@ namespace Server.Controllers.Case.Personal;
 [ApiController]
 [Route("/case/{caseId}")]
 [AuditApi(EventTypeName = "HTTP")]
-public class TabsController : ControllerBase
+public class TabsController(DatabaseContext dbContext, SqidsEncoder<long> sqids, IConfiguration configuration) : ControllerBase
 {
-    private readonly AuditScopeFactory _auditContext;
-    private readonly DatabaseContext _dbContext;
-    private readonly SqidsEncoder<long> _sqids;
-
-    public TabsController(DatabaseContext dbContext, SqidsEncoder<long> sqids)
-    {
-        _dbContext = dbContext;
-        _auditContext = new AuditScopeFactory();
-        _sqids = sqids;
-    }
+    private readonly AuditScopeFactory _auditContext = new();
 
     // GET: /case/?/tabs
     [HttpGet("tabs")]
@@ -29,39 +20,37 @@ public class TabsController : ControllerBase
     [ProducesResponseType(typeof(string), StatusCodes.Status401Unauthorized)]
     [ProducesResponseType(typeof(string), StatusCodes.Status404NotFound)]
     [ProducesResponseType(typeof(string), StatusCodes.Status500InternalServerError)]
-    [Authorize(Roles = "user")]
+    [Authorize]
     public async Task<ActionResult<IEnumerable<API.Tab>>> GetTabs(string caseId)
     {
         // Preflight checks
         PreflightResponse preflightResponse = await PreflightChecks();
 
-        // If preflight checks returned a HTTP error raise it here
+        // If preflight checks returned an HTTP error raise it here
         if (preflightResponse.Error != null) return preflightResponse.Error;
 
-        // If preflight checks details are null return a HTTP 500 error
+        // If preflight checks details are null return an HTTP 500 error
         if (preflightResponse.Details == null)
             return Problem(
                 "Preflight checks failed with an unknown error!"
             );
 
         // Set variables from preflight response
-        string organizationId = preflightResponse.Details.OrganizationId;
-        long userId = preflightResponse.Details.UserId;
+        string emailAddress = preflightResponse.Details.EmailAddress;
         Database.UserSettings userSettings = preflightResponse.Details.UserSettings;
-        long rawCaseId = _sqids.Decode(caseId)[0];
+        long rawCaseId = sqids.Decode(caseId)[0];
 
-        // Log the user's organization ID and the user's ID
+        // Log the user's email address
         IAuditScope auditScope = this.GetCurrentAuditScope();
-        auditScope.SetCustomField("OrganizationID", organizationId);
-        auditScope.SetCustomField("UserID", userId);
+        auditScope.SetCustomField("emailAddress", emailAddress);
 
-        // Get case user from the database including the required entities 
-        Database.CaseUser? caseUser = await _dbContext.CaseUser
-            .Where(cu => cu.Case.Id == rawCaseId && cu.User.Id == userId)
+        // Get case user from the database including the required entities
+        Database.CaseUser? caseUser = await dbContext.CaseUser
+            .Where(cu => cu.Case.Id == rawCaseId && cu.User.EmailAddress == emailAddress)
             .Include(cu => cu.Tabs)
             .SingleOrDefaultAsync();
 
-        // If case user does not exist then return a HTTP 404 error 
+        // If case user does not exist then return an HTTP error404 error
         if (caseUser == null) return NotFound($"The case `{caseId}` does not exist!");
         // The case might not exist or the user does not have access to the case
 
@@ -71,9 +60,7 @@ public class TabsController : ControllerBase
         // Return a list of the user's tabs
         return caseUser.Tabs.Select(t => new API.Tab
         {
-            Id = _sqids.Encode(t.Id),
-            Name = t.Name,
-            Created = TimeZoneInfo.ConvertTimeFromUtc(t.Created, timeZone)
+            Id = sqids.Encode(t.Id), Name = t.Name, Created = TimeZoneInfo.ConvertTimeFromUtc(t.Created, timeZone)
         }).ToList();
     }
 
@@ -84,49 +71,47 @@ public class TabsController : ControllerBase
     [ProducesResponseType(typeof(string), StatusCodes.Status401Unauthorized)]
     [ProducesResponseType(typeof(string), StatusCodes.Status404NotFound)]
     [ProducesResponseType(typeof(string), StatusCodes.Status500InternalServerError)]
-    [Authorize(Roles = "user")]
+    [Authorize]
     public async Task<ActionResult<API.Tab>> GetTab(string caseId, string tabId)
     {
         // Preflight checks
         PreflightResponse preflightResponse = await PreflightChecks();
 
-        // If preflight checks returned a HTTP error raise it here
+        // If preflight checks returned an HTTP error raise it here
         if (preflightResponse.Error != null) return preflightResponse.Error;
 
-        // If preflight checks details are null return a HTTP 500 error
+        // If preflight checks details are null return an HTTP error500 error
         if (preflightResponse.Details == null)
             return Problem(
                 "Preflight checks failed with an unknown error!"
             );
 
         // Set variables from preflight response
-        string organizationId = preflightResponse.Details.OrganizationId;
-        long userId = preflightResponse.Details.UserId;
+        string emailAddress = preflightResponse.Details.EmailAddress;
         Database.UserSettings userSettings = preflightResponse.Details.UserSettings;
-        long rawCaseId = _sqids.Decode(caseId)[0];
+        long rawCaseId = sqids.Decode(caseId)[0];
 
-        // Log the user's organization ID and the user's ID
+        // Log the user's email address
         IAuditScope auditScope = this.GetCurrentAuditScope();
-        auditScope.SetCustomField("OrganizationID", organizationId);
-        auditScope.SetCustomField("UserID", userId);
+        auditScope.SetCustomField("emailAddress", emailAddress);
 
-        // Get case user from the database including the required entities 
-        Database.CaseUser? caseUser = await _dbContext.CaseUser
-            .Where(cu => cu.Case.Id == rawCaseId && cu.User.Id == userId)
+        // Get case user from the database including the required entities
+        Database.CaseUser? caseUser = await dbContext.CaseUser
+            .Where(cu => cu.Case.Id == rawCaseId && cu.User.EmailAddress == emailAddress)
             .Include(cu => cu.Tabs)
             .SingleOrDefaultAsync();
 
-        // If case user does not exist then return a HTTP 404 error 
+        // If case user does not exist then return an HTTP error404 error
         if (caseUser == null) return NotFound($"The case `{caseId}` does not exist!");
         // The case might not exist or the user does not have access to the case
 
-        // Convert tab ID squid to ID 
-        long rawTabId = _sqids.Decode(tabId)[0];
+        // Convert tab ID squid to ID
+        long rawTabId = sqids.Decode(tabId)[0];
 
         // Fetch tab from the database
         Database.Tab? tab = caseUser.Tabs.SingleOrDefault(t => t.Id == rawTabId);
 
-        // If tab is null then return a HTTP 404 error as it does not exist
+        // If tab is null then return an HTTP error404 error as it does not exist
         if (tab == null)
             return NotFound($"A tab with the ID `{tabId}` was not found in the case with the ID `{caseId}`!");
 
@@ -134,12 +119,7 @@ public class TabsController : ControllerBase
         TimeZoneInfo timeZone = TimeZoneInfo.FindSystemTimeZoneById(userSettings.TimeZone);
 
         // Return tab details
-        return new API.Tab
-        {
-            Id = _sqids.Encode(tab.Id),
-            Name = tab.Name,
-            Created = TimeZoneInfo.ConvertTimeFromUtc(tab.Created, timeZone)
-        };
+        return new API.Tab { Id = sqids.Encode(tab.Id), Name = tab.Name, Created = TimeZoneInfo.ConvertTimeFromUtc(tab.Created, timeZone) };
     }
 
     // POST: /case/?/tab
@@ -149,55 +129,51 @@ public class TabsController : ControllerBase
     [ProducesResponseType(typeof(string), StatusCodes.Status401Unauthorized)]
     [ProducesResponseType(typeof(string), StatusCodes.Status404NotFound)]
     [ProducesResponseType(typeof(string), StatusCodes.Status500InternalServerError)]
-    [Authorize(Roles = "user")]
+    [Authorize]
     public async Task<ActionResult<API.Tab>> CreateTab(string caseId, API.AddTab tabAddObject)
     {
         // Preflight checks
         PreflightResponse preflightResponse = await PreflightChecks();
 
-        // If preflight checks returned a HTTP error raise it here
+        // If preflight checks returned an HTTP error raise it here
         if (preflightResponse.Error != null) return preflightResponse.Error;
 
-        // If preflight checks details are null return a HTTP 500 error
+        // If preflight checks details are null return an HTTP error500 error
         if (preflightResponse.Details == null)
             return Problem(
                 "Preflight checks failed with an unknown error!"
             );
 
         // Set variables from preflight response
-        string organizationId = preflightResponse.Details.OrganizationId;
-        long userId = preflightResponse.Details.UserId;
+        string emailAddress = preflightResponse.Details.EmailAddress;
         string userNameJob = preflightResponse.Details.UserNameJob;
         Database.UserSettings userSettings = preflightResponse.Details.UserSettings;
-        long rawCaseId = _sqids.Decode(caseId)[0];
+        long rawCaseId = sqids.Decode(caseId)[0];
 
-        // Log the user's organization ID and the user's ID
+        // Log the user's email adress
         IAuditScope auditScope = this.GetCurrentAuditScope();
-        auditScope.SetCustomField("OrganizationID", organizationId);
-        auditScope.SetCustomField("UserID", userId);
 
-        // Get case user from the database including the required entities 
-        Database.CaseUser? caseUser = await _dbContext.CaseUser
-            .Where(cu => cu.Case.Id == rawCaseId && cu.User.Id == userId)
+        auditScope.SetCustomField("emailAddress", emailAddress);
+
+        // Get case user from the database including the required entities
+        Database.CaseUser? caseUser = await dbContext.CaseUser
+            .Where(cu => cu.Case.Id == rawCaseId && cu.User.EmailAddress == emailAddress)
             .Include(cu => cu.Tabs)
             .Include(cu => cu.Case)
             .SingleOrDefaultAsync();
 
-        // If case user does not exist then return a HTTP 404 error 
+        // If case user does not exist then return an HTTP error404 error
         if (caseUser == null) return NotFound($"The case `{caseId}` does not exist!");
         // The case might not exist or the user does not have access to the case
 
         // Create tab model
-        Database.Tab tabModel = new()
-        {
-            Name = tabAddObject.Name
-        };
+        Database.Tab tabModel = new() { Name = tabAddObject.Name };
 
         // Add tab to database
         caseUser.Tabs.Add(tabModel);
 
         // Save changes to the database
-        await _dbContext.SaveChangesAsync();
+        await dbContext.SaveChangesAsync();
 
         // Log creation of the tab
         await _auditContext.LogAsync("Lighthouse Notes",
@@ -205,7 +181,7 @@ public class TabsController : ControllerBase
             {
                 Action =
                     $"`{userNameJob}` created the tab `{tabModel.Name}` for `{caseUser.Case.DisplayName}`.",
-                UserID = userId, OrganizationID = organizationId
+                EmailAddress = emailAddress
             });
 
         // Get the user's time zone
@@ -215,7 +191,8 @@ public class TabsController : ControllerBase
         return CreatedAtAction(nameof(GetTabs), new { caseId },
             new API.Tab
             {
-                Id = _sqids.Encode(tabModel.Id), Name = tabModel.Name,
+                Id = sqids.Encode(tabModel.Id),
+                Name = tabModel.Name,
                 Created = TimeZoneInfo.ConvertTimeFromUtc(tabModel.Created, timeZone)
             });
     }
@@ -227,45 +204,42 @@ public class TabsController : ControllerBase
     [ProducesResponseType(typeof(string), StatusCodes.Status401Unauthorized)]
     [ProducesResponseType(typeof(string), StatusCodes.Status404NotFound)]
     [ProducesResponseType(typeof(string), StatusCodes.Status500InternalServerError)]
-    [Authorize(Roles = "user")]
+    [Authorize]
     public async Task<ActionResult> GetTabContent(string caseId, string tabId)
     {
         // Preflight checks
         PreflightResponse preflightResponse = await PreflightChecks();
 
-        // If preflight checks returned a HTTP error raise it here
+        // If preflight checks returned an HTTP error raise it here
         if (preflightResponse.Error != null) return preflightResponse.Error;
 
-        // If preflight checks details are null return a HTTP 500 error
+        // If preflight checks details are null return an HTTP error500 error
         if (preflightResponse.Details == null)
             return Problem(
                 "Preflight checks failed with an unknown error!"
             );
 
         // Set variables from preflight response
-        string organizationId = preflightResponse.Details.OrganizationId;
-        Database.OrganizationSettings organizationSettings = preflightResponse.Details.OrganizationSettings;
-        long userId = preflightResponse.Details.UserId;
-        long rawCaseId = _sqids.Decode(caseId)[0];
+        string emailAddress = preflightResponse.Details.EmailAddress;
+        long rawCaseId = sqids.Decode(caseId)[0];
 
-        // Log the user's organization ID and the user's ID
+        // Log the user's email address
         IAuditScope auditScope = this.GetCurrentAuditScope();
-        auditScope.SetCustomField("OrganizationID", organizationId);
-        auditScope.SetCustomField("UserID", userId);
+        auditScope.SetCustomField("emailAddress", emailAddress);
 
-        // Get case user from the database including the required entities 
-        Database.CaseUser? caseUser = await _dbContext.CaseUser
-            .Where(cu => cu.Case.Id == rawCaseId && cu.User.Id == userId)
+        // Get case user from the database including the required entities
+        Database.CaseUser? caseUser = await dbContext.CaseUser
+            .Where(cu => cu.Case.Id == rawCaseId && cu.User.EmailAddress == emailAddress)
             .Include(cu => cu.Tabs)
             .Include(cu => cu.Hashes)
             .SingleOrDefaultAsync();
 
-        // If case user does not exist then return a HTTP 404 error 
-        if (caseUser == null) return NotFound($"The case `{caseId}` does not exist!");
+        // If case user does not exist then return an HTTP error404 error
         // The case might not exist or the user does not have access to the case
+        if (caseUser == null) return NotFound($"The case `{caseId}` does not exist!");
 
-        // Convert tab ID squid to ID 
-        long rawTabId = _sqids.Decode(tabId)[0];
+        // Convert tab ID squid to ID
+        long rawTabId = sqids.Decode(tabId)[0];
 
         // Fetch the tab from the database
         Database.Tab? tab = caseUser.Tabs.SingleOrDefault(t => t.Id == rawTabId);
@@ -276,22 +250,22 @@ public class TabsController : ControllerBase
 
         // Create minio client
         IMinioClient minio = new MinioClient()
-            .WithEndpoint(organizationSettings.S3Endpoint)
-            .WithCredentials(organizationSettings.S3AccessKey, organizationSettings.S3SecretKey)
-            .WithSSL(organizationSettings.S3NetworkEncryption)
+            .WithEndpoint(configuration.GetValue<string>("Minio:Endpoint"))
+            .WithCredentials(configuration.GetValue<string>("Minio:AccessKey"), configuration.GetValue<string>("Minio:SecretKey"))
+            .WithSSL(configuration.GetValue<bool>("Minio:NetworkEncryption"))
             .Build();
 
-        // Create a variable for object path with auth0| removed 
-        string objectPath = $"cases/{caseId}/{_sqids.Encode(userId)}/tabs/{tabId}/content.txt";
+        // Create a variable for object path
+        string objectPath = $"cases/{caseId}/{emailAddress}/tabs/{tabId}/content.txt";
 
         // Check if bucket exists
         bool bucketExists = await minio.BucketExistsAsync(new BucketExistsArgs()
-            .WithBucket(organizationSettings.S3BucketName)
+            .WithBucket(configuration.GetValue<string>("Minio:BucketName"))
         ).ConfigureAwait(false);
 
-        // If bucket does not exist return a HTTP 500 error
+        // If bucket does not exist return an HTTP error500 error
         if (!bucketExists)
-            return Problem($"An S3 Bucket with the name `{organizationSettings.S3BucketName}` does not exist!");
+            return Problem($"An S3 Bucket with the name `{configuration.GetValue<string>("Minio:BucketName")}` does not exist!");
 
         // Create variable to store object metadata
         ObjectStat objectMetadata;
@@ -300,11 +274,11 @@ public class TabsController : ControllerBase
         try
         {
             objectMetadata = await minio.StatObjectAsync(new StatObjectArgs()
-                .WithBucket(organizationSettings.S3BucketName)
+                .WithBucket(configuration.GetValue<string>("Minio:BucketName"))
                 .WithObject(objectPath)
             );
         }
-        // Catch object not found exception and return a HTTP 500 error
+        // Catch object not found exception and return an HTTP error500 error
         catch (ObjectNotFoundException)
         {
             return Problem(
@@ -316,7 +290,7 @@ public class TabsController : ControllerBase
         Database.Hash? objectHashes = caseUser.Hashes.SingleOrDefault(h =>
             h.ObjectName == objectMetadata.ObjectName && h.VersionId == objectMetadata.VersionId);
 
-        // If object hash is null then a hash does not exist so return a HTTP 500 error
+        // If object hash is null then a hash does not exist so return an HTTP error500 error
         if (objectHashes == null)
             return Problem($"Unable to find hash values for the tab with the ID `{tabId}`!",
                 title: "Unable to find hash values for the tab!");
@@ -326,7 +300,7 @@ public class TabsController : ControllerBase
 
         // Get object and copy file contents to stream
         await minio.GetObjectAsync(new GetObjectArgs()
-            .WithBucket(organizationSettings.S3BucketName)
+            .WithBucket(configuration.GetValue<string>("Minio:BucketName"))
             .WithObject(objectPath)
             .WithCallbackStream(stream => { stream.CopyTo(memoryStream); })
         );
@@ -363,79 +337,77 @@ public class TabsController : ControllerBase
     [ProducesResponseType(typeof(string), StatusCodes.Status401Unauthorized)]
     [ProducesResponseType(typeof(string), StatusCodes.Status404NotFound)]
     [ProducesResponseType(typeof(string), StatusCodes.Status500InternalServerError)]
-    [Authorize(Roles = "user")]
-    public async Task<ActionResult> PostTabContent(string caseId, string tabId, [FromForm] IFormFile file)
+    [Authorize]
+    public async Task<ActionResult> PostTabContent(string caseId, string tabId, IFormFile file)
     {
         // Preflight checks
         PreflightResponse preflightResponse = await PreflightChecks();
 
-        // If preflight checks returned a HTTP error raise it here
+        // If preflight checks returned an HTTP error raise it here
         if (preflightResponse.Error != null) return preflightResponse.Error;
 
-        // If preflight checks details are null return a HTTP 500 error
+        // If preflight checks details are null return an HTTP error500 error
         if (preflightResponse.Details == null)
             return Problem(
                 "Preflight checks failed with an unknown error!"
             );
 
         // Set variables from preflight response
-        string organizationId = preflightResponse.Details.OrganizationId;
-        Database.OrganizationSettings organizationSettings = preflightResponse.Details.OrganizationSettings;
-        long userId = preflightResponse.Details.UserId;
+        string emailAddress = preflightResponse.Details.EmailAddress;
         string userNameJob = preflightResponse.Details.UserNameJob;
-        long rawCaseId = _sqids.Decode(caseId)[0];
+        long rawCaseId = sqids.Decode(caseId)[0];
 
-        // Log the user's organization ID and the user's ID
+        // Log the user's email address
         IAuditScope auditScope = this.GetCurrentAuditScope();
-        auditScope.SetCustomField("OrganizationID", organizationId);
-        auditScope.SetCustomField("UserID", userId);
+        auditScope.SetCustomField("emailAddress", emailAddress);
 
-        // Get case user from the database including the required entities 
-        Database.CaseUser? caseUser = await _dbContext.CaseUser
-            .Where(cu => cu.Case.Id == rawCaseId && cu.User.Id == userId)
+        // Get case user from the database including the required entities
+        Database.CaseUser? caseUser = await dbContext.CaseUser
+            .Where(cu => cu.Case.Id == rawCaseId && cu.User.EmailAddress == emailAddress)
             .Include(cu => cu.Tabs)
             .Include(cu => cu.Hashes)
             .Include(cu => cu.Case)
             .SingleOrDefaultAsync();
 
-        // If case user does not exist then return a HTTP 404 error 
-        if (caseUser == null) return NotFound($"The case `{caseId}` does not exist!");
+        // If case user does not exist then return an HTTP 404 error
         // The case might not exist or the user does not have access to the case
+        if (caseUser == null) return NotFound($"The case `{caseId}` does not exist!");
 
-        // Convert tab ID squid to ID 
-        long rawTabId = _sqids.Decode(tabId)[0];
+
+        // Convert tab ID squid to ID
+        long rawTabId = sqids.Decode(tabId)[0];
 
         // Fetch tab from the database
         Database.Tab? tab = caseUser.Tabs.SingleOrDefault(t => t.Id == rawTabId);
 
-        // If tab is null then return a HTTP 404 error
+        // If tab is null then return an HTTP error404 error
         if (tab == null)
             return NotFound($"A tab with the ID `{tabId}` was not found in the case with the ID`{caseId}`!");
 
         // Create minio client
         IMinioClient minio = new MinioClient()
-            .WithEndpoint(organizationSettings.S3Endpoint)
-            .WithCredentials(organizationSettings.S3AccessKey, organizationSettings.S3SecretKey)
-            .WithSSL(organizationSettings.S3NetworkEncryption)
+            .WithEndpoint(configuration.GetValue<string>("Minio:Endpoint"))
+            .WithCredentials(configuration.GetValue<string>("Minio:AccessKey"), configuration.GetValue<string>("Minio:SecretKey"))
+            .WithSSL(configuration.GetValue<bool>("Minio:NetworkEncryption"))
             .Build();
 
-        // Create a variable for filepath with auth0| removed
-        string objectPath = $"cases/{caseId}/{_sqids.Encode(userId)}/tabs/{tabId}/content.txt";
+        // Create a variable for filepath
+        string objectPath = $"cases/{caseId}/{emailAddress}/tabs/{tabId}/content.txt";
 
         // Check if bucket exists
         bool bucketExists = await minio.BucketExistsAsync(new BucketExistsArgs()
-            .WithBucket(organizationSettings.S3BucketName)
+            .WithBucket(configuration.GetValue<string>("Minio:BucketName"))
         ).ConfigureAwait(false);
 
-        // If bucket does not exist return a HTTP 500 error
+        // If bucket does not exist return an HTTP error500 error
         if (!bucketExists)
-            return Problem($"An S3 Bucket with the name `{organizationSettings.S3BucketName}` does not exist!");
+            return Problem($"An S3 Bucket with the name `{configuration.GetValue<string>("Minio:BucketName")}` does not exist!");
 
         // Check if object exists
         try
         {
             await minio.StatObjectAsync(new StatObjectArgs()
-                .WithBucket(organizationSettings.S3BucketName)
+                .WithBucket(configuration.GetValue<string>("Minio:BucketName"))
                 .WithObject(objectPath)
             );
 
@@ -450,7 +422,7 @@ public class TabsController : ControllerBase
 
             // Save the updated file to the s3 bucket
             await minio.PutObjectAsync(new PutObjectArgs()
-                .WithBucket(organizationSettings.S3BucketName)
+                .WithBucket(configuration.GetValue<string>("Minio:BucketName"))
                 .WithObject(objectPath)
                 .WithStreamData(memoryStream)
                 .WithObjectSize(memoryStream.Length)
@@ -462,7 +434,7 @@ public class TabsController : ControllerBase
 
             // Fetch object metadata
             ObjectStat objectMetadata = await minio.StatObjectAsync(new StatObjectArgs()
-                .WithBucket(organizationSettings.S3BucketName)
+                .WithBucket(configuration.GetValue<string>("Minio:BucketName"))
                 .WithObject(objectPath)
             );
 
@@ -484,7 +456,7 @@ public class TabsController : ControllerBase
             });
 
             // Save changes to the database
-            await _dbContext.SaveChangesAsync();
+            await dbContext.SaveChangesAsync();
 
             // Log the addition of content
             await _auditContext.LogAsync("Lighthouse Notes",
@@ -492,7 +464,7 @@ public class TabsController : ControllerBase
                 {
                     Action =
                         $"`{userNameJob}` changed the content in the tab `{tab.Name}` for `{caseUser.Case.DisplayName}`.",
-                    UserID = userId, OrganizationID = organizationId
+                    EmailAddress = emailAddress
                 });
 
             // Return Ok
@@ -512,7 +484,7 @@ public class TabsController : ControllerBase
 
             // Save file to S3 bucket
             await minio.PutObjectAsync(new PutObjectArgs()
-                .WithBucket(organizationSettings.S3BucketName)
+                .WithBucket(configuration.GetValue<string>("Minio:BucketName"))
                 .WithObject(objectPath)
                 .WithStreamData(memoryStream)
                 .WithObjectSize(memoryStream.Length)
@@ -524,7 +496,7 @@ public class TabsController : ControllerBase
 
             // Fetch object metadata
             ObjectStat objectMetadata = await minio.StatObjectAsync(new StatObjectArgs()
-                .WithBucket(organizationSettings.S3BucketName)
+                .WithBucket(configuration.GetValue<string>("Minio:BucketName"))
                 .WithObject(objectPath)
             );
 
@@ -546,7 +518,7 @@ public class TabsController : ControllerBase
             });
 
             // Save changes to the database
-            await _dbContext.SaveChangesAsync();
+            await dbContext.SaveChangesAsync();
 
             // Log the creation
             await _auditContext.LogAsync("Lighthouse Notes",
@@ -554,7 +526,7 @@ public class TabsController : ControllerBase
                 {
                     Action =
                         $"`{userNameJob}` created content in the tab `{tab.Name}` for `{caseUser.Case.DisplayName}`.",
-                    UserID = userId, OrganizationID = organizationId
+                    EmailAddress = emailAddress
                 });
 
             // Return Ok
@@ -569,46 +541,30 @@ public class TabsController : ControllerBase
 
     private async Task<PreflightResponse> PreflightChecks()
     {
-        // Get user ID from claim
-        string? auth0UserId = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value;
+        // Get user email from JWT claim
+        string? emailAddress = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Email)?.Value;
 
-        // If user ID is null then it does not exist in JWT so return a HTTP 400 error
-        if (auth0UserId == null)
-            return new PreflightResponse
-                { Error = BadRequest("User ID can not be found in the JSON Web Token (JWT)!") };
+        // If user email is null then it does not exist in JWT so return an HTTP 400 error
+        if (string.IsNullOrEmpty(emailAddress))
+            return new PreflightResponse { Error = BadRequest("Email Address cannot be found in the JSON Web Token (JWT)!") };
 
-        // Get organization ID from claim
-        string? organizationId = User.Claims.FirstOrDefault(c => c.Type == "org_id")?.Value;
-
-        // If organization ID  is null then it does not exist in JWT so return a HTTP 400 error
-        if (organizationId == null)
-            return new PreflightResponse
-                { Error = BadRequest("Organization ID can not be found in the JSON Web Token (JWT)!") };
-
-        // Select organization ID, organization settings, user ID and user name and job and settings from the user table
-        PreflightResponseDetails? userQueryResult = await _dbContext.User
-            .Where(u => u.Auth0Id == auth0UserId && u.Organization.Id == organizationId)
+        // Query user details including roles and application settings
+        PreflightResponseDetails? preflightData = await dbContext.User
+            .Where(u => u.EmailAddress == emailAddress)
+            .Include(u => u.Settings)
             .Select(u => new PreflightResponseDetails
             {
-                OrganizationId = u.Organization.Id,
-                OrganizationSettings = u.Organization.Settings,
-                UserId = u.Id,
-                UserNameJob = $"{u.DisplayName} ({u.JobTitle})",
-                UserSettings = u.Settings
-            }).SingleOrDefaultAsync();
+                EmailAddress = u.EmailAddress, UserNameJob = $"{u.DisplayName} ({u.JobTitle})", UserSettings = u.Settings
+            })
+            .SingleOrDefaultAsync();
 
-        // If query result is null then the user does not exit in the organization so return a HTTP 404 error
-        if (userQueryResult == null)
-            return new PreflightResponse
-            {
-                Error = NotFound(
-                    $"A user with the Auth0 user ID `{auth0UserId}` was not found in the organization with the Auth0 organization ID `{organizationId}`!")
-            };
+        // If query result is null then the user does not exist
+        if (preflightData == null)
+            return new PreflightResponse { Error = NotFound($"A user with the user email: `{emailAddress}` was not found!") }
+                ;
 
-        return new PreflightResponse
-        {
-            Details = userQueryResult
-        };
+        // Return preflight response
+        return new PreflightResponse { Details = preflightData };
     }
 
     private class PreflightResponse
@@ -619,9 +575,7 @@ public class TabsController : ControllerBase
 
     private class PreflightResponseDetails
     {
-        public required string OrganizationId { get; init; }
-        public required Database.OrganizationSettings OrganizationSettings { get; init; }
-        public long UserId { get; init; }
+        public required string EmailAddress { get; init; }
         public required string UserNameJob { get; init; }
         public required Database.UserSettings UserSettings { get; init; }
     }
