@@ -2,12 +2,20 @@
 
 namespace Server.Data;
 
-public class DatabaseContext : DbContext
+public class DatabaseContext(DbContextOptions<DatabaseContext> options) : DbContext(options)
 {
-    public DatabaseContext(DbContextOptions<DatabaseContext> options)
-        : base(options)
-    {
-    }
+    // Tables
+    public DbSet<Database.Event> Event => Set<Database.Event>();
+    public DbSet<Database.User> User => Set<Database.User>();
+    public IEnumerable<Database.UserSettings> UserSettings => Set<Database.UserSettings>();
+    public DbSet<Database.Case> Case => Set<Database.Case>();
+    public DbSet<Database.CaseUser> CaseUser => Set<Database.CaseUser>();
+    public IEnumerable<Database.ContemporaneousNote> ContemporaneousNote => Set<Database.ContemporaneousNote>();
+    public IEnumerable<Database.Tab> Tab => Set<Database.Tab>();
+    public DbSet<Database.SharedTab> SharedTab => Set<Database.SharedTab>();
+    public IEnumerable<Database.Exhibit> Exhibit => Set<Database.Exhibit>();
+
+    public IEnumerable<Database.Hash> Hash => Set<Database.Hash>();
 
     // Override SaveChanges function and call OnBeforeSaving (Chestnut, 2019)
     public override int SaveChanges(bool acceptAllChangesOnSuccess)
@@ -17,14 +25,10 @@ public class DatabaseContext : DbContext
     }
 
     // Override SaveChangesAsync function and call OnBeforeSaving (Chestnut, 2019)
-    public override async Task<int> SaveChangesAsync(
-        bool acceptAllChangesOnSuccess,
-        CancellationToken cancellationToken = default
-    )
+    public override async Task<int> SaveChangesAsync(bool acceptAllChangesOnSuccess, CancellationToken cancellationToken = default)
     {
         OnBeforeSaving();
-        return await base.SaveChangesAsync(acceptAllChangesOnSuccess,
-            cancellationToken);
+        return await base.SaveChangesAsync(acceptAllChangesOnSuccess, cancellationToken);
     }
 
     // Before saving changes to the Database Update or Set Modified / Created time (Chestnut, 2019)
@@ -34,75 +38,54 @@ public class DatabaseContext : DbContext
         DateTime utcNow = DateTime.UtcNow;
 
         foreach (EntityEntry entry in entries)
-            // for entities that inherit from BaseEntity,
-            // set UpdatedOn / CreatedOn appropriately
+            // For entities that inherit from BaseEntity, set UpdatedOn / CreatedOn appropriately
+        {
             if (entry.Entity is Database.Base trackable)
                 // ReSharper disable once SwitchStatementMissingSomeEnumCasesNoDefault
+            {
                 switch (entry.State)
                 {
                     case EntityState.Modified:
-                        // set the updated date to "now"
+                        // Set the updated date to "now"
                         trackable.Modified = utcNow;
 
-                        // mark property as "don't touch"
-                        // we don't want to update on a Modify operation
+                        // Mark property as "don't touch" we don't want to update on a Modify operation
                         entry.Property("Created").IsModified = false;
                         break;
 
                     case EntityState.Added:
-                        // set both updated and created date to "now"
+                        // Set both updated and created date to "now"
                         trackable.Created = utcNow;
                         trackable.Modified = utcNow;
                         break;
                 }
+            }
+        }
     }
 
-    public DbSet<Database.Event> Event => Set<Database.Event>();
-    public DbSet<Database.Organization> Organization => Set<Database.Organization>();
-
-    public IEnumerable<Database.OrganizationSettings> OrganizationSettings =>
-        Set<Database.OrganizationSettings>();
-
-    public DbSet<Database.User> User => Set<Database.User>();
-
-    public IEnumerable<Database.UserSettings> UserSettings =>
-        Set<Database.UserSettings>();
-
-    public IEnumerable<Database.Role> Role => Set<Database.Role>();
-    public DbSet<Database.Case> Case => Set<Database.Case>();
-    public DbSet<Database.CaseUser> CaseUser => Set<Database.CaseUser>();
-    public IEnumerable<Database.ContemporaneousNote> ContemporaneousNote => Set<Database.ContemporaneousNote>();
-    public IEnumerable<Database.Tab> Tab => Set<Database.Tab>();
-    public DbSet<Database.SharedTab> SharedTab => Set<Database.SharedTab>();
-    public IEnumerable<Database.Exhibit> Exhibit => Set<Database.Exhibit>();
-    public IEnumerable<Database.Hash> Hash => Set<Database.Hash>();
-
+    // Model Creation - Create relationships
     protected override void OnModelCreating(ModelBuilder modelBuilder)
     {
-        // Organization Configuration Relationship (one to one)
-        modelBuilder.Entity<Database.Organization>()
-            .HasOne(e => e.Settings)
-            .WithOne(e => e.Organization)
-            .HasForeignKey<Database.OrganizationSettings>("OrganizationId");
-
-        // User settings relationship (one to one) 
+        // User settings relationship (one to one)
         modelBuilder.Entity<Database.User>()
             .HasOne(e => e.Settings)
             .WithOne(e => e.User)
-            .HasForeignKey<Database.UserSettings>("UserId");
+            .HasForeignKey<Database.UserSettings>("EmailAddress");
 
         // User events relationship (one to many)
         modelBuilder.Entity<Database.User>()
             .HasMany(u => u.Events)
             .WithOne(e => e.User)
-            .HasForeignKey("UserId")
+            .HasForeignKey("EmailAddress")
             .IsRequired(false);
 
+        // Event created default value is SQL now()
         modelBuilder
             .Entity<Database.Event>()
             .Property(e => e.Created)
             .HasDefaultValueSql("now()");
 
+        // Event updated default value is SQL now()
         modelBuilder
             .Entity<Database.Event>()
             .Property(e => e.Updated)
