@@ -6,48 +6,35 @@ using Minio.DataModel;
 using Minio.DataModel.Args;
 using Minio.Exceptions;
 
-namespace Server.Export;
+namespace Server.Data;
 
-public class Functions
+public class Functions(
+    SqidsEncoder<long> sqids,
+    string caseId,
+    string emailAddress,
+    Database.Case sCase,
+    Database.CaseUser sCaseUser,
+    TimeZoneInfo timeZone,
+    IConfiguration configuration)
 {
-    private readonly SqidsEncoder<long> _sqids;
-    private string _caseId;
-    private string _userId;
-    private Database.Case _case;
-    private Database.CaseUser _caseUser;
-    private Database.OrganizationSettings _organizationSettings;
-    private TimeZoneInfo _timeZone;
-
-    public Functions(SqidsEncoder<long> sqids, string caseId, string userId, Database.Case sCase,
-        Database.CaseUser caseUser, Database.OrganizationSettings organizationSettings, TimeZoneInfo timeZone)
-    {
-        _sqids = sqids;
-        _caseId = caseId;
-        _userId = userId;
-        _case = sCase;
-        _caseUser = caseUser;
-        _organizationSettings = organizationSettings;
-        _timeZone = timeZone;
-    }
-
     public async Task<API.ContemporaneousNotesExport> GetContemporaneousNote(
         Database.ContemporaneousNote contemporaneousNote)
     {
         // Create minio client
         IMinioClient minio = new MinioClient()
-            .WithEndpoint(_organizationSettings.S3Endpoint)
-            .WithCredentials(_organizationSettings.S3AccessKey, _organizationSettings.S3SecretKey)
-            .WithSSL(_organizationSettings.S3NetworkEncryption)
+            .WithEndpoint(configuration.GetValue<string>("Minio:Endpoint"))
+            .WithCredentials(configuration.GetValue<string>("Minio:AccessKey"), configuration.GetValue<string>("Minio:SecretKey"))
+            .WithSSL(configuration.GetValue<bool>("Minio:NetworkEncryption"))
             .Build();
 
-        // Variables 
+        // Variables
         MemoryStream memoryStream = new();
         using MD5 md5 = MD5.Create();
         using SHA256 sha256 = SHA256.Create();
 
-        // Create a variable for object path with auth0| removed 
+        // Create a variable for object path
         string objectPath =
-            $"cases/{_caseId}/{_userId}/contemporaneous-notes/{_sqids.Encode(contemporaneousNote.Id)}/note.txt";
+            $"cases/{caseId}/{emailAddress}/contemporaneous-notes/{sqids.Encode(contemporaneousNote.Id)}/note.txt";
 
         // Create variable to store object metadata
         ObjectStat objectMetadata;
@@ -56,23 +43,23 @@ public class Functions
         try
         {
             objectMetadata = await minio.StatObjectAsync(new StatObjectArgs()
-                .WithBucket(_organizationSettings.S3BucketName)
+                .WithBucket(configuration.GetValue<string>("Minio:BucketName"))
                 .WithObject(objectPath)
             );
         }
         // Catch object not found exception and return HTTP 500 error
         catch (ObjectNotFoundException)
         {
-            return new API.ContemporaneousNotesExport()
+            return new API.ContemporaneousNotesExport
             {
                 Content =
-                    $"Can not find the S3 object for contemporaneous note: `{_sqids.Encode(contemporaneousNote.Id)}` at the following path: `{objectPath}`.",
-                DateTime = TimeZoneInfo.ConvertTimeFromUtc(contemporaneousNote.Created, _timeZone)
+                    $"Can not find the S3 object for contemporaneous note: `{sqids.Encode(contemporaneousNote.Id)}` at the following path: `{objectPath}`.",
+                DateTime = TimeZoneInfo.ConvertTimeFromUtc(contemporaneousNote.Created, timeZone)
             };
         }
 
         // Fetch object hash from database
-        Database.Hash? objectHashes = _caseUser.Hashes.SingleOrDefault(h =>
+        Database.Hash? objectHashes = sCaseUser.Hashes.SingleOrDefault(h =>
             h.ObjectName == objectMetadata.ObjectName && h.VersionId == objectMetadata.VersionId);
 
         // If object hash is null then a hash does not exist so return HTTP 500 error
@@ -80,13 +67,13 @@ public class Functions
             return new API.ContemporaneousNotesExport
             {
                 Content =
-                    $"Unable to find hash value for contemporaneous note with the ID `{_sqids.Encode(contemporaneousNote.Id)}` at the path `{objectPath}`!",
-                DateTime = TimeZoneInfo.ConvertTimeFromUtc(contemporaneousNote.Created, _timeZone)
+                    $"Unable to find hash value for contemporaneous note with the ID `{sqids.Encode(contemporaneousNote.Id)}` at the path `{objectPath}`!",
+                DateTime = TimeZoneInfo.ConvertTimeFromUtc(contemporaneousNote.Created, timeZone)
             };
 
         // Get object and copy file contents to stream
         await minio.GetObjectAsync(new GetObjectArgs()
-            .WithBucket(_organizationSettings.S3BucketName)
+            .WithBucket(configuration.GetValue<string>("Minio:BucketName"))
             .WithObject(objectPath)
             .WithCallbackStream(stream => { stream.CopyTo(memoryStream); })
         );
@@ -103,17 +90,17 @@ public class Functions
             return new API.ContemporaneousNotesExport
             {
                 Content =
-                    $"MD5 hash verification failed for contemporaneous note with the ID `{_sqids.Encode(contemporaneousNote.Id)}` at the path `{objectPath}`!",
-                DateTime = TimeZoneInfo.ConvertTimeFromUtc(contemporaneousNote.Created, _timeZone)
+                    $"MD5 hash verification failed for contemporaneous note with the ID `{sqids.Encode(contemporaneousNote.Id)}` at the path `{objectPath}`!",
+                DateTime = TimeZoneInfo.ConvertTimeFromUtc(contemporaneousNote.Created, timeZone)
             };
 
         // Check SHA256 hash generated matches hash in the database
         if (BitConverter.ToString(sha256Hash).Replace("-", "").ToLowerInvariant() != objectHashes.ShaHash)
-            return new API.ContemporaneousNotesExport()
+            return new API.ContemporaneousNotesExport
             {
                 Content =
-                    $"SHA256 hash verification failed for contemporaneous note with the ID `{_sqids.Encode(contemporaneousNote.Id)}` at the path `{objectPath}`!",
-                DateTime = TimeZoneInfo.ConvertTimeFromUtc(contemporaneousNote.Created, _timeZone)
+                    $"SHA256 hash verification failed for contemporaneous note with the ID `{sqids.Encode(contemporaneousNote.Id)}` at the path `{objectPath}`!",
+                DateTime = TimeZoneInfo.ConvertTimeFromUtc(contemporaneousNote.Created, timeZone)
             };
 
         // Set memory stream position to 0 as per github.com/minio/minio/issues/6274
@@ -138,7 +125,7 @@ public class Functions
                 // Create variable with file name
                 string fileName = img.Attributes["src"].Value.Replace(".path/", "");
                 string imageObjectPath =
-                    $"cases/{_caseId}/{_userId}/contemporaneous-notes/images/{HttpUtility.UrlDecode(fileName)}";
+                    $"cases/{caseId}/{emailAddress}/contemporaneous-notes/images/{HttpUtility.UrlDecode(fileName)}";
 
                 // Get the image path
                 string imagePath = await GetImage(imageObjectPath);
@@ -149,8 +136,7 @@ public class Functions
 
         return new API.ContemporaneousNotesExport
         {
-            Content = htmlDoc.DocumentNode.OuterHtml,
-            DateTime = TimeZoneInfo.ConvertTimeFromUtc(contemporaneousNote.Created, _timeZone)
+            Content = htmlDoc.DocumentNode.OuterHtml, DateTime = TimeZoneInfo.ConvertTimeFromUtc(contemporaneousNote.Created, timeZone)
         };
     }
 
@@ -158,13 +144,13 @@ public class Functions
     {
         // Create minio client
         IMinioClient minio = new MinioClient()
-            .WithEndpoint(_organizationSettings.S3Endpoint)
-            .WithCredentials(_organizationSettings.S3AccessKey, _organizationSettings.S3SecretKey)
-            .WithSSL(_organizationSettings.S3NetworkEncryption)
+            .WithEndpoint(configuration.GetValue<string>("Minio:Endpoint"))
+            .WithCredentials(configuration.GetValue<string>("Minio:AccessKey"), configuration.GetValue<string>("Minio:SecretKey"))
+            .WithSSL(configuration.GetValue<bool>("Minio:NetworkEncryption"))
             .Build();
 
-        // Create a variable for object path with auth0| removed 
-        string objectPath = $"cases/{_caseId}/{_userId}/tabs/{_sqids.Encode(tab.Id)}/content.txt";
+        // Create a variable for object path
+        string objectPath = $"cases/{caseId}/{emailAddress}/tabs/{sqids.Encode(tab.Id)}/content.txt";
 
         // Create variable to store object metadata
         ObjectStat objectMetadata;
@@ -173,31 +159,30 @@ public class Functions
         try
         {
             objectMetadata = await minio.StatObjectAsync(new StatObjectArgs()
-                .WithBucket(_organizationSettings.S3BucketName)
+                .WithBucket(configuration.GetValue<string>("Minio:BucketName"))
                 .WithObject(objectPath)
             );
         }
-        // Catch object not found exception and return a HTTP 500 error
+        // Catch object not found exception and return an HTTP 500 error
         catch (ObjectNotFoundException)
         {
-            return new API.TabExport()
+            return new API.TabExport
             {
                 Content =
-                    $"Can not find the S3 object for the tab with the ID `{_sqids.Encode(tab.Id)}` at the following path `{objectPath}`.",
+                    $"Can not find the S3 object for the tab with the ID `{sqids.Encode(tab.Id)}` at the following path `{objectPath}`.",
                 Name = tab.Name
             };
         }
 
         // Fetch object hash from database
-        Database.Hash? objectHashes = _caseUser.Hashes.SingleOrDefault(h =>
+        Database.Hash? objectHashes = sCaseUser.Hashes.SingleOrDefault(h =>
             h.ObjectName == objectMetadata.ObjectName && h.VersionId == objectMetadata.VersionId);
 
-        // If object hash is null then a hash does not exist so return a HTTP 500 error
+        // If object hash is null then a hash does not exist so return am HTTP 500 error
         if (objectHashes == null)
-            return new API.TabExport()
+            return new API.TabExport
             {
-                Content = $"Unable to find hash values for the tab with the ID `{_sqids.Encode(tab.Id)}`!",
-                Name = tab.Name
+                Content = $"Unable to find hash values for the tab with the ID `{sqids.Encode(tab.Id)}`!", Name = tab.Name
             };
 
         // Create memory stream to store file contents
@@ -205,7 +190,7 @@ public class Functions
 
         // Get object and copy file contents to stream
         await minio.GetObjectAsync(new GetObjectArgs()
-            .WithBucket(_organizationSettings.S3BucketName)
+            .WithBucket(configuration.GetValue<string>("Minio:BucketName"))
             .WithObject(objectPath)
             .WithCallbackStream(stream => { stream.CopyTo(memoryStream); })
         );
@@ -223,18 +208,18 @@ public class Functions
 
         // Check generated MD5 hash matches the hash in the database
         if (BitConverter.ToString(md5Hash).Replace("-", "").ToLowerInvariant() != objectHashes.Md5Hash)
-            return new API.TabExport()
+            return new API.TabExport
             {
                 Content =
-                    $"MD5 hash verification failed for tab with the ID `{_sqids.Encode(tab.Id)}` at the path `{objectPath}`!",
+                    $"MD5 hash verification failed for tab with the ID `{sqids.Encode(tab.Id)}` at the path `{objectPath}`!",
                 Name = tab.Name
             };
         // Check generated SHA256 hash matches the hash in the database
         if (BitConverter.ToString(sha256Hash).Replace("-", "").ToLowerInvariant() != objectHashes.ShaHash)
-            return new API.TabExport()
+            return new API.TabExport
             {
                 Content =
-                    $"SHA256 hash verification failed for tab with the ID `{_sqids.Encode(tab.Id)}` at the path `{objectPath}`!",
+                    $"SHA256 hash verification failed for tab with the ID `{sqids.Encode(tab.Id)}` at the path `{objectPath}`!",
                 Name = tab.Name
             };
 
@@ -260,8 +245,8 @@ public class Functions
                 // Create variable with file name
                 string fileName = img.Attributes["src"].Value.Replace(".path/", "");
 
-                // Get presigned s3 url and update image src 
-                string imageObjectPath = $"cases/{_caseId}/{_userId}/tabs/images/{HttpUtility.UrlDecode(fileName)}";
+                // Get presigned s3 url and update image src
+                string imageObjectPath = $"cases/{caseId}/{emailAddress}/tabs/images/{HttpUtility.UrlDecode(fileName)}";
 
                 // Get the image path
                 string imagePath = await GetImage(imageObjectPath);
@@ -270,11 +255,7 @@ public class Functions
                 img.Attributes["src"].Value = imagePath;
             }
 
-        return new API.TabExport()
-        {
-            Name = tab.Name,
-            Content = htmlDoc.DocumentNode.OuterHtml
-        };
+        return new API.TabExport { Name = tab.Name, Content = htmlDoc.DocumentNode.OuterHtml };
     }
 
     public async Task<API.SharedContemporaneousNotesExport> GetSharedContemporaneousNote(
@@ -282,19 +263,19 @@ public class Functions
     {
         // Create minio client
         IMinioClient minio = new MinioClient()
-            .WithEndpoint(_organizationSettings.S3Endpoint)
-            .WithCredentials(_organizationSettings.S3AccessKey, _organizationSettings.S3SecretKey)
-            .WithSSL(_organizationSettings.S3NetworkEncryption)
+            .WithEndpoint(configuration.GetValue<string>("Minio:Endpoint"))
+            .WithCredentials(configuration.GetValue<string>("Minio:AccessKey"), configuration.GetValue<string>("Minio:SecretKey"))
+            .WithSSL(configuration.GetValue<bool>("Minio:NetworkEncryption"))
             .Build();
 
-        // Variables 
+        // Variables
         MemoryStream memoryStream = new();
         using MD5 md5 = MD5.Create();
         using SHA256 sha256 = SHA256.Create();
 
-        // Create a variable for object path with auth0| removed 
+        // Create a variable for object path
         string objectPath =
-            $"cases/{_caseId}/shared/contemporaneous-notes/{_sqids.Encode(contemporaneousNote.Id)}/note.txt";
+            $"cases/{caseId}/shared/contemporaneous-notes/{sqids.Encode(contemporaneousNote.Id)}/note.txt";
 
         // Create variable to store object metadata
         ObjectStat objectMetadata;
@@ -303,68 +284,53 @@ public class Functions
         try
         {
             objectMetadata = await minio.StatObjectAsync(new StatObjectArgs()
-                .WithBucket(_organizationSettings.S3BucketName)
+                .WithBucket(configuration.GetValue<string>("Minio:BucketName"))
                 .WithObject(objectPath)
             );
         }
         // Catch object not found exception and return HTTP 500 error
         catch (ObjectNotFoundException)
         {
-            return new API.SharedContemporaneousNotesExport()
+            return new API.SharedContemporaneousNotesExport
             {
                 Content =
-                    $"Can not find the S3 object for the shared contemporaneous note: `{_sqids.Encode(contemporaneousNote.Id)}` at the following path: `{objectPath}`.",
-
-                Created = TimeZoneInfo.ConvertTimeFromUtc(contemporaneousNote.Created, _timeZone),
+                    $"Can not find the S3 object for the shared contemporaneous note: `{sqids.Encode(contemporaneousNote.Id)}` at the following path: `{objectPath}`.",
+                Created = TimeZoneInfo.ConvertTimeFromUtc(contemporaneousNote.Created, timeZone),
                 Creator = new API.User
                 {
-                    Id = _sqids.Encode(contemporaneousNote.Creator.Id), Auth0Id = contemporaneousNote.Creator.Auth0Id,
+                    EmailAddress = contemporaneousNote.Creator.EmailAddress,
                     JobTitle = contemporaneousNote.Creator.JobTitle,
                     DisplayName = contemporaneousNote.Creator.DisplayName,
-                    GivenName = contemporaneousNote.Creator.GivenName, LastName = contemporaneousNote.Creator.LastName,
-                    EmailAddress = contemporaneousNote.Creator.EmailAddress,
-                    ProfilePicture = contemporaneousNote.Creator.ProfilePicture,
-                    Organization = new API.Organization
-                    {
-                        Name = contemporaneousNote.Creator.Organization.DisplayName,
-                        DisplayName = contemporaneousNote.Creator.Organization.DisplayName
-                    },
-                    Roles = contemporaneousNote.Creator.Roles.Select(r => r.Name).ToList()
+                    GivenName = contemporaneousNote.Creator.GivenName,
+                    LastName = contemporaneousNote.Creator.LastName
                 }
             };
         }
 
         // Fetch object hash from database
-        Database.SharedHash? objectHashes = _case.SharedHashes.SingleOrDefault(h =>
+        Database.SharedHash? objectHashes = sCase.SharedHashes.SingleOrDefault(h =>
             h.ObjectName == objectMetadata.ObjectName && h.VersionId == objectMetadata.VersionId);
 
         // If object hash is null then a hash does not exist so return HTTP 500 error
         if (objectHashes == null)
-            return new API.SharedContemporaneousNotesExport()
+            return new API.SharedContemporaneousNotesExport
             {
                 Content =
-                    $"Unable to find hash value for the shared contemporaneous note with the ID `{_sqids.Encode(contemporaneousNote.Id)}` at the path `{objectPath}`!",
-                Created = TimeZoneInfo.ConvertTimeFromUtc(contemporaneousNote.Created, _timeZone),
+                    $"Unable to find hash value for the shared contemporaneous note with the ID `{sqids.Encode(contemporaneousNote.Id)}` at the path `{objectPath}`!",
+                Created = TimeZoneInfo.ConvertTimeFromUtc(contemporaneousNote.Created, timeZone),
                 Creator = new API.User
                 {
-                    Id = _sqids.Encode(contemporaneousNote.Creator.Id), Auth0Id = contemporaneousNote.Creator.Auth0Id,
+                    EmailAddress = contemporaneousNote.Creator.EmailAddress,
                     JobTitle = contemporaneousNote.Creator.JobTitle,
                     DisplayName = contemporaneousNote.Creator.DisplayName,
-                    GivenName = contemporaneousNote.Creator.GivenName, LastName = contemporaneousNote.Creator.LastName,
-                    EmailAddress = contemporaneousNote.Creator.EmailAddress,
-                    ProfilePicture = contemporaneousNote.Creator.ProfilePicture,
-                    Organization = new API.Organization
-                    {
-                        Name = contemporaneousNote.Creator.Organization.DisplayName,
-                        DisplayName = contemporaneousNote.Creator.Organization.DisplayName
-                    },
-                    Roles = contemporaneousNote.Creator.Roles.Select(r => r.Name).ToList()
+                    GivenName = contemporaneousNote.Creator.GivenName,
+                    LastName = contemporaneousNote.Creator.LastName
                 }
             };
 
         // Get object and copy file contents to stream
         await minio.GetObjectAsync(new GetObjectArgs()
-            .WithBucket(_organizationSettings.S3BucketName)
+            .WithBucket(configuration.GetValue<string>("Minio:BucketName"))
             .WithObject(objectPath)
             .WithCallbackStream(stream => { stream.CopyTo(memoryStream); })
         );
@@ -378,49 +344,35 @@ public class Functions
 
         // Check MD5 hash generated matches hash in the database
         if (BitConverter.ToString(md5Hash).Replace("-", "").ToLowerInvariant() != objectHashes.Md5Hash)
-            return new API.SharedContemporaneousNotesExport()
+            return new API.SharedContemporaneousNotesExport
             {
                 Content =
-                    $"MD5 hash verification failed for contemporaneous note with the ID `{_sqids.Encode(contemporaneousNote.Id)}` at the path `{objectPath}`!",
-                Created = TimeZoneInfo.ConvertTimeFromUtc(contemporaneousNote.Created, _timeZone),
+                    $"MD5 hash verification failed for contemporaneous note with the ID `{sqids.Encode(contemporaneousNote.Id)}` at the path `{objectPath}`!",
+                Created = TimeZoneInfo.ConvertTimeFromUtc(contemporaneousNote.Created, timeZone),
                 Creator = new API.User
                 {
-                    Id = _sqids.Encode(contemporaneousNote.Creator.Id), Auth0Id = contemporaneousNote.Creator.Auth0Id,
+                    EmailAddress = contemporaneousNote.Creator.EmailAddress,
                     JobTitle = contemporaneousNote.Creator.JobTitle,
                     DisplayName = contemporaneousNote.Creator.DisplayName,
-                    GivenName = contemporaneousNote.Creator.GivenName, LastName = contemporaneousNote.Creator.LastName,
-                    EmailAddress = contemporaneousNote.Creator.EmailAddress,
-                    ProfilePicture = contemporaneousNote.Creator.ProfilePicture,
-                    Organization = new API.Organization
-                    {
-                        Name = contemporaneousNote.Creator.Organization.DisplayName,
-                        DisplayName = contemporaneousNote.Creator.Organization.DisplayName
-                    },
-                    Roles = contemporaneousNote.Creator.Roles.Select(r => r.Name).ToList()
+                    GivenName = contemporaneousNote.Creator.GivenName,
+                    LastName = contemporaneousNote.Creator.LastName
                 }
             };
 
         // Check SHA256 hash generated matches hash in the database
         if (BitConverter.ToString(sha256Hash).Replace("-", "").ToLowerInvariant() != objectHashes.ShaHash)
-            return new API.SharedContemporaneousNotesExport()
+            return new API.SharedContemporaneousNotesExport
             {
                 Content =
-                    $"SHA256 hash verification failed for contemporaneous note with the ID `{_sqids.Encode(contemporaneousNote.Id)}` at the path `{objectPath}`!",
-                Created = TimeZoneInfo.ConvertTimeFromUtc(contemporaneousNote.Created, _timeZone),
+                    $"SHA256 hash verification failed for contemporaneous note with the ID `{sqids.Encode(contemporaneousNote.Id)}` at the path `{objectPath}`!",
+                Created = TimeZoneInfo.ConvertTimeFromUtc(contemporaneousNote.Created, timeZone),
                 Creator = new API.User
                 {
-                    Id = _sqids.Encode(contemporaneousNote.Creator.Id), Auth0Id = contemporaneousNote.Creator.Auth0Id,
+                    EmailAddress = contemporaneousNote.Creator.EmailAddress,
                     JobTitle = contemporaneousNote.Creator.JobTitle,
                     DisplayName = contemporaneousNote.Creator.DisplayName,
-                    GivenName = contemporaneousNote.Creator.GivenName, LastName = contemporaneousNote.Creator.LastName,
-                    EmailAddress = contemporaneousNote.Creator.EmailAddress,
-                    ProfilePicture = contemporaneousNote.Creator.ProfilePicture,
-                    Organization = new API.Organization
-                    {
-                        Name = contemporaneousNote.Creator.Organization.DisplayName,
-                        DisplayName = contemporaneousNote.Creator.Organization.DisplayName
-                    },
-                    Roles = contemporaneousNote.Creator.Roles.Select(r => r.Name).ToList()
+                    GivenName = contemporaneousNote.Creator.GivenName,
+                    LastName = contemporaneousNote.Creator.LastName
                 }
             };
 
@@ -446,7 +398,7 @@ public class Functions
                 // Create variable with file name
                 string fileName = img.Attributes["src"].Value.Replace(".path/", "");
                 string imageObjectPath =
-                    $"cases/{_caseId}/shared/contemporaneous-notes/images/{HttpUtility.UrlDecode(fileName)}";
+                    $"cases/{caseId}/shared/contemporaneous-notes/images/{HttpUtility.UrlDecode(fileName)}";
 
                 // Get the image path
                 string imagePath = await GetSharedImage(imageObjectPath);
@@ -455,24 +407,17 @@ public class Functions
                 img.Attributes["src"].Value = imagePath;
             }
 
-        return new API.SharedContemporaneousNotesExport()
+        return new API.SharedContemporaneousNotesExport
         {
             Content = htmlDoc.DocumentNode.OuterHtml,
-            Created = TimeZoneInfo.ConvertTimeFromUtc(contemporaneousNote.Created, _timeZone),
+            Created = TimeZoneInfo.ConvertTimeFromUtc(contemporaneousNote.Created, timeZone),
             Creator = new API.User
             {
-                Id = _sqids.Encode(contemporaneousNote.Creator.Id), Auth0Id = contemporaneousNote.Creator.Auth0Id,
+                EmailAddress = contemporaneousNote.Creator.EmailAddress,
                 JobTitle = contemporaneousNote.Creator.JobTitle,
                 DisplayName = contemporaneousNote.Creator.DisplayName,
-                GivenName = contemporaneousNote.Creator.GivenName, LastName = contemporaneousNote.Creator.LastName,
-                EmailAddress = contemporaneousNote.Creator.EmailAddress,
-                ProfilePicture = contemporaneousNote.Creator.ProfilePicture,
-                Organization = new API.Organization
-                {
-                    Name = contemporaneousNote.Creator.Organization.DisplayName,
-                    DisplayName = contemporaneousNote.Creator.Organization.DisplayName
-                },
-                Roles = contemporaneousNote.Creator.Roles.Select(r => r.Name).ToList()
+                GivenName = contemporaneousNote.Creator.GivenName,
+                LastName = contemporaneousNote.Creator.LastName
             }
         };
     }
@@ -481,13 +426,13 @@ public class Functions
     {
         // Create minio client
         IMinioClient minio = new MinioClient()
-            .WithEndpoint(_organizationSettings.S3Endpoint)
-            .WithCredentials(_organizationSettings.S3AccessKey, _organizationSettings.S3SecretKey)
-            .WithSSL(_organizationSettings.S3NetworkEncryption)
+            .WithEndpoint(configuration.GetValue<string>("Minio:Endpoint"))
+            .WithCredentials(configuration.GetValue<string>("Minio:AccessKey"), configuration.GetValue<string>("Minio:SecretKey"))
+            .WithSSL(configuration.GetValue<bool>("Minio:NetworkEncryption"))
             .Build();
 
-        // Create a variable for object path with auth0| removed 
-        string objectPath = $"cases/{_caseId}/shared/tabs/{_sqids.Encode(tab.Id)}/content.txt";
+        // Create a variable for object path
+        string objectPath = $"cases/{caseId}/shared/tabs/{sqids.Encode(tab.Id)}/content.txt";
 
         // Create variable to store object metadata
         ObjectStat objectMetadata;
@@ -496,59 +441,47 @@ public class Functions
         try
         {
             objectMetadata = await minio.StatObjectAsync(new StatObjectArgs()
-                .WithBucket(_organizationSettings.S3BucketName)
+                .WithBucket(configuration.GetValue<string>("Minio:BucketName"))
                 .WithObject(objectPath)
             );
         }
-        // Catch object not found exception and return a HTTP 500 error
+        // Catch object not found exception and return an HTTP 500 error
         catch (ObjectNotFoundException)
         {
             return new API.SharedTabExport
             {
                 Content =
-                    $"Can not find the S3 object for the tab with the ID `{_sqids.Encode(tab.Id)}` at the following path `{objectPath}`.",
+                    $"Can not find the S3 object for the tab with the ID `{sqids.Encode(tab.Id)}` at the following path `{objectPath}`.",
                 Name = tab.Name,
-                Created = TimeZoneInfo.ConvertTimeFromUtc(tab.Created, _timeZone),
+                Created = TimeZoneInfo.ConvertTimeFromUtc(tab.Created, timeZone),
                 Creator = new API.User
                 {
-                    Id = _sqids.Encode(tab.Creator.Id), Auth0Id = tab.Creator.Auth0Id,
+                    EmailAddress = tab.Creator.EmailAddress,
                     JobTitle = tab.Creator.JobTitle,
                     DisplayName = tab.Creator.DisplayName,
-                    GivenName = tab.Creator.GivenName, LastName = tab.Creator.LastName,
-                    EmailAddress = tab.Creator.EmailAddress,
-                    ProfilePicture = tab.Creator.ProfilePicture,
-                    Organization = new API.Organization
-                    {
-                        Name = tab.Creator.Organization.DisplayName, DisplayName = tab.Creator.Organization.DisplayName
-                    },
-                    Roles = tab.Creator.Roles.Select(r => r.Name).ToList()
+                    GivenName = tab.Creator.GivenName,
+                    LastName = tab.Creator.LastName
                 }
             };
         }
 
-        Database.SharedHash? objectHashes = _case.SharedHashes.SingleOrDefault(h =>
+        Database.SharedHash? objectHashes = sCase.SharedHashes.SingleOrDefault(h =>
             h.ObjectName == objectMetadata.ObjectName && h.VersionId == objectMetadata.VersionId);
 
-        // If object hash is null then a hash does not exist so return a HTTP 500 error
+        // If object hash is null then a hash does not exist so return an HTTP 500 error
         if (objectHashes == null)
             return new API.SharedTabExport
             {
-                Content = $"Unable to find hash values for the tab with the ID `{_sqids.Encode(tab.Id)}`!",
+                Content = $"Unable to find hash values for the tab with the ID `{sqids.Encode(tab.Id)}`!",
                 Name = tab.Name,
-                Created = TimeZoneInfo.ConvertTimeFromUtc(tab.Created, _timeZone),
+                Created = TimeZoneInfo.ConvertTimeFromUtc(tab.Created, timeZone),
                 Creator = new API.User
                 {
-                    Id = _sqids.Encode(tab.Creator.Id), Auth0Id = tab.Creator.Auth0Id,
                     JobTitle = tab.Creator.JobTitle,
                     DisplayName = tab.Creator.DisplayName,
-                    GivenName = tab.Creator.GivenName, LastName = tab.Creator.LastName,
-                    EmailAddress = tab.Creator.EmailAddress,
-                    ProfilePicture = tab.Creator.ProfilePicture,
-                    Organization = new API.Organization
-                    {
-                        Name = tab.Creator.Organization.DisplayName, DisplayName = tab.Creator.Organization.DisplayName
-                    },
-                    Roles = tab.Creator.Roles.Select(r => r.Name).ToList()
+                    GivenName = tab.Creator.GivenName,
+                    LastName = tab.Creator.LastName,
+                    EmailAddress = tab.Creator.EmailAddress
                 }
             };
 
@@ -557,7 +490,7 @@ public class Functions
 
         // Get object and copy file contents to stream
         await minio.GetObjectAsync(new GetObjectArgs()
-            .WithBucket(_organizationSettings.S3BucketName)
+            .WithBucket(configuration.GetValue<string>("Minio:BucketName"))
             .WithObject(objectPath)
             .WithCallbackStream(stream => { stream.CopyTo(memoryStream); })
         );
@@ -578,22 +511,16 @@ public class Functions
             return new API.SharedTabExport
             {
                 Content =
-                    $"MD5 hash verification failed for tab with the ID `{_sqids.Encode(tab.Id)}` at the path `{objectPath}`!",
+                    $"MD5 hash verification failed for tab with the ID `{sqids.Encode(tab.Id)}` at the path `{objectPath}`!",
                 Name = tab.Name,
-                Created = TimeZoneInfo.ConvertTimeFromUtc(tab.Created, _timeZone),
+                Created = TimeZoneInfo.ConvertTimeFromUtc(tab.Created, timeZone),
                 Creator = new API.User
                 {
-                    Id = _sqids.Encode(tab.Creator.Id), Auth0Id = tab.Creator.Auth0Id,
+                    EmailAddress = tab.Creator.EmailAddress,
                     JobTitle = tab.Creator.JobTitle,
                     DisplayName = tab.Creator.DisplayName,
-                    GivenName = tab.Creator.GivenName, LastName = tab.Creator.LastName,
-                    EmailAddress = tab.Creator.EmailAddress,
-                    ProfilePicture = tab.Creator.ProfilePicture,
-                    Organization = new API.Organization
-                    {
-                        Name = tab.Creator.Organization.DisplayName, DisplayName = tab.Creator.Organization.DisplayName
-                    },
-                    Roles = tab.Creator.Roles.Select(r => r.Name).ToList()
+                    GivenName = tab.Creator.GivenName,
+                    LastName = tab.Creator.LastName
                 }
             };
         // Check generated SHA256 hash matches the hash in the database
@@ -601,22 +528,16 @@ public class Functions
             return new API.SharedTabExport
             {
                 Content =
-                    $"SHA256 hash verification failed for tab with the ID `{_sqids.Encode(tab.Id)}` at the path `{objectPath}`!",
+                    $"SHA256 hash verification failed for tab with the ID `{sqids.Encode(tab.Id)}` at the path `{objectPath}`!",
                 Name = tab.Name,
-                Created = TimeZoneInfo.ConvertTimeFromUtc(tab.Created, _timeZone),
+                Created = TimeZoneInfo.ConvertTimeFromUtc(tab.Created, timeZone),
                 Creator = new API.User
                 {
-                    Id = _sqids.Encode(tab.Creator.Id), Auth0Id = tab.Creator.Auth0Id,
+                    EmailAddress = tab.Creator.EmailAddress,
                     JobTitle = tab.Creator.JobTitle,
                     DisplayName = tab.Creator.DisplayName,
-                    GivenName = tab.Creator.GivenName, LastName = tab.Creator.LastName,
-                    EmailAddress = tab.Creator.EmailAddress,
-                    ProfilePicture = tab.Creator.ProfilePicture,
-                    Organization = new API.Organization
-                    {
-                        Name = tab.Creator.Organization.DisplayName, DisplayName = tab.Creator.Organization.DisplayName
-                    },
-                    Roles = tab.Creator.Roles.Select(r => r.Name).ToList()
+                    GivenName = tab.Creator.GivenName,
+                    LastName = tab.Creator.LastName
                 }
             };
 
@@ -642,8 +563,8 @@ public class Functions
                 // Create variable with file name
                 string fileName = img.Attributes["src"].Value.Replace(".path/", "");
 
-                // Get presigned s3 url and update image src 
-                string imageObjectPath = $"cases/{_caseId}/shared/tabs/images/{HttpUtility.UrlDecode(fileName)}";
+                // Get presigned s3 url and update image src
+                string imageObjectPath = $"cases/{caseId}/shared/tabs/images/{HttpUtility.UrlDecode(fileName)}";
 
                 // Get the image path
                 string imagePath = await GetSharedImage(imageObjectPath);
@@ -656,20 +577,14 @@ public class Functions
         {
             Name = tab.Name,
             Content = htmlDoc.DocumentNode.OuterHtml,
-            Created = TimeZoneInfo.ConvertTimeFromUtc(tab.Created, _timeZone),
+            Created = TimeZoneInfo.ConvertTimeFromUtc(tab.Created, timeZone),
             Creator = new API.User
             {
-                Id = _sqids.Encode(tab.Creator.Id), Auth0Id = tab.Creator.Auth0Id,
+                EmailAddress = tab.Creator.EmailAddress,
                 JobTitle = tab.Creator.JobTitle,
                 DisplayName = tab.Creator.DisplayName,
-                GivenName = tab.Creator.GivenName, LastName = tab.Creator.LastName,
-                EmailAddress = tab.Creator.EmailAddress,
-                ProfilePicture = tab.Creator.ProfilePicture,
-                Organization = new API.Organization
-                {
-                    Name = tab.Creator.Organization.DisplayName, DisplayName = tab.Creator.Organization.DisplayName
-                },
-                Roles = tab.Creator.Roles.Select(r => r.Name).ToList()
+                GivenName = tab.Creator.GivenName,
+                LastName = tab.Creator.LastName
             }
         };
     }
@@ -678,9 +593,9 @@ public class Functions
     {
         // Create minio client
         IMinioClient minio = new MinioClient()
-            .WithEndpoint(_organizationSettings.S3Endpoint)
-            .WithCredentials(_organizationSettings.S3AccessKey, _organizationSettings.S3SecretKey)
-            .WithSSL(_organizationSettings.S3NetworkEncryption)
+            .WithEndpoint(configuration.GetValue<string>("Minio:Endpoint"))
+            .WithCredentials(configuration.GetValue<string>("Minio:AccessKey"), configuration.GetValue<string>("Minio:SecretKey"))
+            .WithSSL(configuration.GetValue<bool>("Minio:NetworkEncryption"))
             .Build();
 
         ObjectStat? objectMetadata;
@@ -688,21 +603,21 @@ public class Functions
         try
         {
             objectMetadata = await minio.StatObjectAsync(new StatObjectArgs()
-                .WithBucket(_organizationSettings.S3BucketName)
+                .WithBucket(configuration.GetValue<string>("Minio:BucketName"))
                 .WithObject(objectPath)
             );
         }
-        // If object does not exist return a HTTP 404 error
+        // If object does not exist return an HTTP 404 error
         catch (ObjectNotFoundException)
         {
             return "./Export/image-error.jpeg";
         }
 
         // Fetch object hash from database
-        Database.Hash? objectHashes = _caseUser.Hashes.SingleOrDefault(h =>
+        Database.Hash? objectHashes = sCaseUser.Hashes.SingleOrDefault(h =>
             h.ObjectName == objectMetadata.ObjectName && h.VersionId == objectMetadata.VersionId);
 
-        // If object hash is null then a hash does not exist so return a HTTP 500 error
+        // If object hash is null then a hash does not exist so return an HTTP 500 error
         if (objectHashes == null) return "./Export/image-error.jpeg";
 
         // Create memory stream to store file contents
@@ -710,7 +625,7 @@ public class Functions
 
         // Get object and copy file contents to stream
         await minio.GetObjectAsync(new GetObjectArgs()
-            .WithBucket(_organizationSettings.S3BucketName)
+            .WithBucket(configuration.GetValue<string>("Minio:BucketName"))
             .WithObject(objectPath)
             .WithCallbackStream(stream => { stream.CopyTo(memoryStream); })
         );
@@ -727,16 +642,16 @@ public class Functions
         byte[] sha256Hash = await sha256.ComputeHashAsync(memoryStream);
 
         // Check generated MD5 hash matches the hash in the database
-        if (BitConverter.ToString(md5Hash).Replace("-", "").ToLowerInvariant() != objectHashes.Md5Hash)
+        if (!BitConverter.ToString(md5Hash).Replace("-", "").Equals(objectHashes.Md5Hash, StringComparison.InvariantCultureIgnoreCase))
             return "./Export/image-error.jpeg";
 
         // Check generated SHA256 hash matches the hash in the database
-        if (BitConverter.ToString(sha256Hash).Replace("-", "").ToLowerInvariant() != objectHashes.ShaHash)
+        if (!BitConverter.ToString(sha256Hash).Replace("-", "").Equals(objectHashes.ShaHash, StringComparison.InvariantCultureIgnoreCase))
             return "./Export/image-error.jpeg";
 
-        // Fetch presigned url for object  
+        // Fetch presigned url for object
         string presignedS3Url = await minio.PresignedGetObjectAsync(new PresignedGetObjectArgs()
-            .WithBucket(_organizationSettings.S3BucketName)
+            .WithBucket(configuration.GetValue<string>("Minio:BucketName"))
             .WithObject(objectPath)
             .WithExpiry(3600)
         );
@@ -748,9 +663,9 @@ public class Functions
     {
         // Create minio client
         IMinioClient minio = new MinioClient()
-            .WithEndpoint(_organizationSettings.S3Endpoint)
-            .WithCredentials(_organizationSettings.S3AccessKey, _organizationSettings.S3SecretKey)
-            .WithSSL(_organizationSettings.S3NetworkEncryption)
+            .WithEndpoint(configuration.GetValue<string>("Minio:Endpoint"))
+            .WithCredentials(configuration.GetValue<string>("Minio:AccessKey"), configuration.GetValue<string>("Minio:SecretKey"))
+            .WithSSL(configuration.GetValue<bool>("Minio:NetworkEncryption"))
             .Build();
 
         ObjectStat? objectMetadata;
@@ -758,21 +673,21 @@ public class Functions
         try
         {
             objectMetadata = await minio.StatObjectAsync(new StatObjectArgs()
-                .WithBucket(_organizationSettings.S3BucketName)
+                .WithBucket(configuration.GetValue<string>("Minio:BucketName"))
                 .WithObject(objectPath)
             );
         }
-        // If object does not exist return a HTTP 404 error
+        // If object does not exist return an HTTP 404 error
         catch (ObjectNotFoundException)
         {
             return "./Export/image-error.jpeg";
         }
 
         // Fetch object hash from database
-        Database.SharedHash? objectHashes = _case.SharedHashes.SingleOrDefault(h =>
+        Database.SharedHash? objectHashes = sCase.SharedHashes.SingleOrDefault(h =>
             h.ObjectName == objectMetadata.ObjectName && h.VersionId == objectMetadata.VersionId);
 
-        // If object hash is null then a hash does not exist so return a HTTP 500 error
+        // If object hash is null then a hash does not exist so return an HTTP 500 error
         if (objectHashes == null) return "./Export/image-error.jpeg";
 
         // Create memory stream to store file contents
@@ -780,7 +695,7 @@ public class Functions
 
         // Get object and copy file contents to stream
         await minio.GetObjectAsync(new GetObjectArgs()
-            .WithBucket(_organizationSettings.S3BucketName)
+            .WithBucket(configuration.GetValue<string>("Minio:BucketName"))
             .WithObject(objectPath)
             .WithCallbackStream(stream => { stream.CopyTo(memoryStream); })
         );
@@ -797,16 +712,16 @@ public class Functions
         byte[] sha256Hash = await sha256.ComputeHashAsync(memoryStream);
 
         // Check generated MD5 hash matches the hash in the database
-        if (BitConverter.ToString(md5Hash).Replace("-", "").ToLowerInvariant() != objectHashes.Md5Hash)
+        if (!BitConverter.ToString(md5Hash).Replace("-", "").Equals(objectHashes.Md5Hash, StringComparison.InvariantCultureIgnoreCase))
             return "./Export/image-error.jpeg";
 
         // Check generated SHA256 hash matches the hash in the database
-        if (BitConverter.ToString(sha256Hash).Replace("-", "").ToLowerInvariant() != objectHashes.ShaHash)
+        if (!BitConverter.ToString(sha256Hash).Replace("-", "").Equals(objectHashes.ShaHash, StringComparison.InvariantCultureIgnoreCase))
             return "./Export/image-error.jpeg";
 
-        // Fetch presigned url for object  
+        // Fetch presigned url for object
         string presignedS3Url = await minio.PresignedGetObjectAsync(new PresignedGetObjectArgs()
-            .WithBucket(_organizationSettings.S3BucketName)
+            .WithBucket(configuration.GetValue<string>("Minio:BucketName"))
             .WithObject(objectPath)
             .WithExpiry(3600)
         );
