@@ -6,7 +6,7 @@ public class ContemporaneousNote
 {
     public long Id { get; init; }
     public long CaseId { get; set; }
-    public long UserId { get; set; }
+    public string? EmailAddress { get; set; }
     public string? Content { get; set; }
 }
 
