@@ -1,8 +1,10 @@
+// ReSharper disable UnusedAutoPropertyAccessor.Global
+
 namespace Server.Models.Search;
 
 public class User
 {
-    public long Id { get; set; }
+    public string? EmailAddress { get; init; }
     public string? JobTitle { get; init; }
     public string? DisplayName { get; init; }
     public string? GivenName { get; init; }
